### 2019-03-02 #287 @JingningShi
Submission：MtreeRing
---

Submitting Author: Jingning Shi (@JingningShi )

Repository:   https://github.com/JingningShi/MtreeRing
Version submitted:   1.2
Version accepted: 1.3

---

* Paste the full DESCRIPTION file inside a code block below:

```
Package: MtreeRing
Title: A Shiny Application for Automatic Measurements of Tree-Ring Widths on Digital Images
Version: 1.2
Description: Use morphological image processing and edge detection algorithms to
automatically measure tree ring widths on digital images. Users can also manually mark
tree rings on species with complex anatomical structures. The arcs of inner-rings and
angles of successive inclined ring boundaries are used to correct ring-width series. The
package provides a Shiny-based application, allowing R beginners to easily analyze tree
ring images and export ring-width series in standard file formats.
Authors@R: c(person("Jingning", "Shi", role = c("aut", "cre"),
                    email = "snow940220@bjfu.edu.cn"),
             person("Wei", "Xiang", role = "aut"))
    knitr
    R (>= 3.3.0),
    magrittr (>= 1.5)
    png,
    jpeg,
    tiff,
    bmp,
    magick,
    imager,
    dplR,
    spatstat,
    measuRing,
    shiny,
    shinydashboard,
    shinyWidgets

NeedsCompilation: no
Repository: CRAN
```

* Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below. If you are unsure, we suggest you make a pre-submission inquiry.):

* Explain how and why the package falls under these categories (briefly, 1-2 sentences):

The package aims to enable R users to obtain data from scanned tree-ring  images.

* Who is the target audience and what are scientific applications of this package?

Anyone interested in measuring ring-width series from increment cores or stem disks, for example,  meteorologists and ecologists. Scientific applications include accurate measurements of ring-width series and data correction.

* Are there other R packages that accomplish the same thing? If so, how does

To our knowledge, the package [measuRing](https://cran.r-project.org/web/packages/measuRing/index.html) can also be used for measuring tree ring widths on scanned images.  Compared to measuRing, we provide two additional approaches (watershed segmentation and Canny edge detector) for the automatic identification of ring borders.  Furthermore, our package provides a Shiny-based app. This beginner-friendly app allows interactive viewing and manipulation of tree ring images and data files.

* If you made a pre-submission enquiry, please paste the link to the corresponding issue, forum post, or other discussion, or @tag the editor you contacted.

It's already there.
