+++ .travis.yml
@@ -5,11 +5,7 @@ r:
   - oldrel
   - release
   - devel
-before_install:
-  - sudo apt-get install -y libmagick++-dev
-  - sudo add-apt-repository -y ppa:opencpu/imagemagick
-  - sudo apt-get update
-  - sudo apt-get install -y libmagick++-dev             
+       
 r_packages:
   - covr
 addons:
