+++ .travis.yml
@@ -4,8 +4,11 @@ language: r
 r:
   - oldrel
   - release
-  - devel
-       
+  - devel 
+before_install:
+  - sudo apt-get update
+  - sudo apt-get install -y libmagick++-dev 
+  - sudo add-apt-repository -y ppa:opencpu/imagemagick         
 r_packages:
   - covr
 addons:
