+++ .travis.yml
@@ -2,7 +2,9 @@
 
 language: r
 r:
+  - oldrel
   - release
+  - devel 
 before_install:
   - sudo apt-get install -y libmagick++-dev
   - sudo add-apt-repository -y ppa:opencpu/imagemagick
