+++ .travis.yml
@@ -10,10 +10,14 @@ before_install:
   - sudo add-apt-repository -y ppa:opencpu/imagemagick
   - sudo apt-get update
   - sudo apt-get install -y libmagick++-dev             
+r_check_args: '--ignore-vignettes'
+r_build_args: '--no-build-vignettes'
 r_packages:
   - covr
 addons:
   apt:
+    sources:
+      - sourceline: 'ppa:opencpu/imagemagick'
     packages:
       - libtiff-dev
       - libpng-dev
