+++ .travis.yml
@@ -1,10 +1,9 @@
+# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
+
 language: r
 r:
-  - oldrel
   - release
   - devel 
-r_check_args: '--ignore-vignettes'
-r_build_args: '--no-build-vignettes'
 before_install:
   - sudo apt-get install -y libmagick++-dev
   - sudo add-apt-repository -y ppa:opencpu/imagemagick
