+++ README.Rmd
@@ -1,182 +1,0 @@
-
-MtreeRing
-=======
-
-<!--require(knitr);require(markdown);knit("README.Rmd")-->
-```{r echo=FALSE}
-knitr::opts_chunk$set(
-  comment = "#>",
-  collapse = TRUE,
-  warning = FALSE,
-  message = FALSE,
-  out.width = "100%"
-)
-```
-
-**Authors:** [Jingning Shi](https://www.researchgate.net/profile/Jingning_Shi), [Wei Xiang](https://www.researchgate.net/profile/Wei_Xiang15)<br/>
-**License:** [GPL3](https://cran.r-project.org/web/licenses/GPL-3)
-
-<!--我的徽章-->
-[![Build Status](https://travis-ci.org/JingningShi/MtreeRing.svg?branch=master)](https://travis-ci.org/JingningShi/MtreeRing.svg?branch=master)
-[![cran checks](https://cranchecks.info/badges/worst/MtreeRing)](https://cranchecks.info/pkgs/MtreeRing)
-[![cran version](https://www.r-pkg.org/badges/version/MtreeRing)](https://cran.r-project.org/package=MtreeRing)
-[![Downloads](https://cranlogs.r-pkg.org/badges/MtreeRing)](https://CRAN.R-project.org/package=MtreeRing)
-[![Total Downloads](https://cranlogs.r-pkg.org/badges/grand-total/MtreeRing?color=orange)](https://CRAN.R-project.org/package=MtreeRing)
-<!--[![codecov](https://codecov.io/gh/JingningShi/MtreeRing/branch/master/graphs/badge.svg)](https://codecov.io/gh/JingningShi/MtreeRing)-->
-
-
-`MtreeRing` is a tool for automatically measuring tree-ring width using image processing techniques.
-
-## Installation
-
-Install the stable version from CRAN
-
-```{r eval=FALSE}
-install.packages("MtreeRing")
-```
-
-or the development version from GitHub
-
-```{r eval=FALSE}
-# install.packages("devtools")
-devtools::install_github("JingningShi/MtreeRing")
-```
-
-## Utilities
-
-```{r eval=FALSE}
-library(MtreeRing)
-```
-
-### Read an image
-
-```{r eval=FALSE}
-## Read and plot a tree ring image
-img.name <- system.file("001.png", package = "MtreeRing")
-t1 <- imgInput(img = img.name, dpi = 1200)
-```
-
-### Detect ring borders 
-
-After plotting the image, the automatic detection of ring borders can be performed using three alternative methods: (1) watershed algorithm; (2) Canny edge detector; (3) a linear detection algorithm from R package [measuRing](https://CRAN.R-project.org/package=measuRing).
-
-```{r eval=FALSE}
-## Split a long core sample into 2 pieces to
-## get better display performance and use the
-## watershed algorithm to detect ring borders:
-t2 <- autoDetect(ring.data = t1, seg = 2, method = 'watershed')
-```
-
-<center><img src="https://github.com/JingningShi/GifRepo/blob/master/figures/README-img001.png" width = 65% height = 65% /></center>
-<center>Figure 1. The automatic detection of ring borders</center>
-
-## Shiny application
-
-```{r eval=FALSE}
-MtreeRing::launchMtRApp()
-```
-
-This command allows to run a Shiny-based application within the system's default web browser. The app provides a beginner-friendly graphical interface and supports more flexible mouse-based interactions.
-
-The dashboard has three components: a header, sidebar and body, like this
-
-<img src="https://github.com/JingningShi/GifRepo/blob/master/figures/README-img002.png" width = 85% height = 85% />  
-
-## A simple workflow for Shiny app
-
-### 1. Image upload
-
[…]
