+++ tests/testthat/test-calcRingWidth.R
@@ -5,13 +5,15 @@ test_that("calcRingWidth returns a data.frame", {
   path1 <- system.file("001.png", package = "MtreeRing")
   img1 <- imgInput(img = path1, dpi = 1200)
   
-  t1 <- autoDetect(img1, auto.path = T, method = 'watershed', incline = T)
+  t1 <- autoDetect(ring.data = img1, auto.path = T,
+                   method = 'watershed', incline = T)
   dev.off()
-  t2 <- autoDetect(img1, auto.path = T, method = 'canny', incline = F)
+  t2 <- autoDetect(ring.data = img1, auto.path = T, 
+                   method = 'canny', incline = F)
   dev.off()
   
-  rw1 <- calcRingWidth(t1, '940220')
-  rw2 <- calcRingWidth(t2, '940220')
+  rw1 <- calcRingWidth(ring.data = t1, '940220')
+  rw2 <- calcRingWidth(ring.data = t2, '940220')
   
   expect_is(rw1, "data.frame")
   expect_is(rw2, "data.frame")
