+++ tests/testthat/test-imgInput.R
@@ -10,15 +10,15 @@ test_that("imgInput plots a tree ring image and returns a magick object", {
   
   ## Read and plot the image:
   img1 <- imgInput(img = path1, dpi = 1200)
-  dev.off()
+  dev.off(attributes(img1)$dn)
   img2 <- imgInput(img = path2, dpi = 1200)
-  dev.off()
+  dev.off(attributes(img2)$dn)
   img3 <- imgInput(img = path3, dpi = 1200)
-  dev.off()
+  dev.off(attributes(img3)$dn)
   img4 <- imgInput(img = path4, dpi = 1200)
-  dev.off()
+  dev.off(attributes(img4)$dn)
   img5 <- imgInput(img = path5, dpi = 1200)
-  dev.off()
+  dev.off(attributes(img5)$dn)
   
   expect_is(img1, "magick-image")
   expect_is(img2, "magick-image")
