### 2019-03-17 #287 @annakrystalli
---

Hello @JingningShi. 👋

 Many thanks for your submission and apologies for the delay in completing the editors checks for you!

Firstly, it looks like a really cool package 😎.  There's a few issues I found during the editor's checks though that will need addressing before we proceed:

- [ ] We require that rOpenSci packages are [documented using Roxygen](https://ropensci.github.io/dev_guide/building.html#documentation). That means including documentation with the functions in the `R/` directory and automatically generating the `NAMESPACE` and contents of the `man/` folder using `devtools::document()`. At the minute there are no roxygen comments with the functions and the `NAMESPACE` and `man/` files appears to have been created manually. Please consult [our guidance on documentation](https://ropensci.github.io/dev_guide/building.html#documentation) or the chapter on [Object documentation](http://r-pkgs.had.co.nz/man.html) for more information.

- [ ] We require that [at least 75% code coverage by tests](https://ropensci.github.io/dev_guide/building.html#testing). Is there a reason coverage is so low (total coverage 37%) and that there are files that are not covered at all? These files especially:

    ```
    R/createServer.R: 0.00%
    R/nearPith.R: 0.00%
    R/autoDetect.R: 61.68%
    R/visualSelect.R: 66.67%
    ```

- [ ] Running the tests throws up these warnings for me. Any ideas:
    ```
    test-imgInput.R:14: warning: imgInput plots a tree ring image and returns a magick object
    TIFFFetchNormalTag: Incompatible type for "RichTIFFIPTC"; tag ignored

    test-imgInput.R:16: warning: imgInput plots a tree ring image and returns a magick object
    more bytes available than required in image
    ```
- [ ] Incomplete final line found on `'/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/DESCRIPTION'`. This means the DESCRIPTION file needs a blank line at the end of it. This enables functionality which depends on parsing the `DESCRIPTION` file correctly to work. There are a number of other files that need similar attention (see results of `goodpractice::gp()` checks beflow)

- [ ] `.travis.yml` needs to be added to `.buildignore` so it is ignored when package is being built
    * use `usethis::use_build_ignore(".travis.yml")`

### goodpractice

As mentioned above, these files need at blank line at the end of each.

```
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/autoDetect.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/calcRingWidth.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/correct.color.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/createServer.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/createUI.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/f.img.middle.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/f.morphological.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/f.plot.double.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/f.plot.single.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/imgInput.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/nearPith.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/split.label.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/visualSelect.R'
incomplete final line found on '/Users/Anna/Documents/workflows/rOpenSci/editorials/MtreeRing/R/water.im.R'
```

Also, it's also good practice to:

- [ ]  ✖ write short and simple functions. These functions have
    high cyclomatic complexity:
    - createServer (254),
    - autoDetect (52).
- [ ]  ✖ do not use "Depends" in DESCRIPTION, as it can cause name
    clashes, and poor interaction with other packages. Use "Imports"
    instead.
- [ ]  ✖ avoid sapply(), it is not type safe. It might return a
- [ ]  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.

Let me know when these issues have been addressed and I will begin looking for reviewers. Also feel free to reach out if have any more questions
