+++ NAMESPACE
@@ -1,26 +1,58 @@
-# MtreeRing NAMESPACE file
+# Generated by roxygen2: do not edit by hand
 
-import(shiny, shinydashboard, shinyWidgets)
-importFrom(imager,as.cimg,cannyEdges,erode_rect,dilate_rect,mclosing_square,
-           imgradient,watershed,threshold)
-importFrom(magick,image_crop,image_info,image_read,image_convert,image_rotate,image_resize)
+export(autoDetect)
+export(calcRingWidth)
+export(imgInput)
+export(launchMtRApp)
+export(nearPith)
+export(visualSelect)
 import(dplyr)
-importFrom(png,readPNG)
-importFrom(jpeg,readJPEG)
-importFrom(tiff,readTIFF)
+import(shiny)
+import(shinyWidgets)
+import(shinydashboard)
 importFrom(bmp,read.bmp)
 importFrom(dplR,write.rwl)
-importFrom(spatstat,connected,im)
+importFrom(grDevices,as.raster)
+importFrom(grDevices,dev.cur)
+importFrom(grDevices,dev.list)
+importFrom(grDevices,dev.new)
+importFrom(grDevices,dev.off)
+importFrom(grDevices,dev.set)
+importFrom(graphics,abline)
+importFrom(graphics,axis)
+importFrom(graphics,layout)
+importFrom(graphics,lines)
+importFrom(graphics,locator)
+importFrom(graphics,mtext)
+importFrom(graphics,par)
+importFrom(graphics,plot)
+importFrom(graphics,points)
+importFrom(graphics,rasterImage)
+importFrom(graphics,segments)
+importFrom(graphics,text)
+importFrom(graphics,title)
+importFrom(imager,as.cimg)
+importFrom(imager,cannyEdges)
+importFrom(imager,dilate_rect)
+importFrom(imager,erode_rect)
+importFrom(imager,imgradient)
+importFrom(imager,mclosing_square)
+importFrom(imager,threshold)
+importFrom(imager,watershed)
+importFrom(jpeg,readJPEG)
+importFrom(magick,image_convert)
+importFrom(magick,image_crop)
+importFrom(magick,image_info)
+importFrom(magick,image_read)
+importFrom(magick,image_resize)
+importFrom(magick,image_rotate)
 importFrom(magrittr,"%>%")
-importFrom(measuRing,graySmoothed,linearDetect)
-
-importFrom("grDevices", "as.raster", "dev.cur", "dev.new", "dev.set",
-		   "dev.off", "dev.list")
-importFrom("graphics", "abline", "axis", "layout", "lines", "locator",
-		 "mtext", "par", "plot", "points", "rasterImage", "segments",
-		 "text", "title")
-importFrom("utils", "write.csv")
-importFrom("stats", "coef", "lm")
-
-## Export function names
-export(imgInput,autoDetect,launchMtRApp,calcRingWidth,visualSelect)
+importFrom(measuRing,graySmoothed)
+importFrom(measuRing,linearDetect)
+importFrom(png,readPNG)
+importFrom(spatstat,connected)
+importFrom(spatstat,im)
+importFrom(stats,coef)
+importFrom(stats,lm)
+importFrom(tiff,readTIFF)
+importFrom(utils,write.csv)
