+++ DESCRIPTION
@@ -2,20 +2,18 @@ Package: MtreeRing
 Type: Package
 Title: A Shiny Application for Automatic Measurements of Tree-Ring Widths on Digital Images
 Version: 1.2
-Authors@R: c(person("Jingning", "Shi", role = c("aut", "cre"),
-                    email = "snow940220@bjfu.edu.cn"),
-             person("Wei", "Xiang", role = "aut"))
+Authors@R: c(person("Jingning", "Shi", role = c("aut", "cre"), email = "snow940220@bjfu.edu.cn"), person("Wei", "Xiang", role = "aut"))
 Author: Jingning Shi [aut, cre], Wei Xiang [aut]
 Maintainer: Jingning Shi <snow940220@bjfu.edu.cn>
 Suggests: testthat, knitr
-Depends: R (>= 3.3.0), magrittr (>= 1.5)
-Imports: png, jpeg, tiff, bmp, magick, imager, dplR, spatstat, measuRing, shiny, dplyr, shinydashboard, shinyWidgets
+Imports: magrittr (>= 1.5), png, jpeg, tiff, bmp, magick, imager, dplR, spatstat, measuRing, shiny, dplyr, shinydashboard, shinyWidgets
 Description: Use morphological image processing and edge detection algorithms to automatically measure tree ring widths on digital images. Users can also manually mark tree rings on species with complex anatomical structures. The arcs of inner-rings and angles of successive inclined ring boundaries are used to correct ring-width series. The package provides a Shiny-based application, allowing R beginners to easily analyze tree ring images and export ring-width series in standard file formats.
 License: GPL-3
 LazyData: TRUE
 NeedsCompilation: no
 Repository: CRAN
 Encoding: UTF-8
-VignetteBuilder: knitr
 URL: https://github.com/JingningShi/MtreeRing
 BugReports: https://github.com/JingningShi/MtreeRing/issues
+VignetteBuilder: knitr
+RoxygenNote: 6.1.1
