+++ man/visualSelect.Rd
@@ -1,36 +1,56 @@
-\encoding{UTF-8}
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/visualSelect.R
 \name{visualSelect}
 \alias{visualSelect}
-
 \title{Edit ring borders visually}
-\description{This function can delete existing ring borders and add new borders visually.}
 \usage{
-visualSelect(ring.data, del = NULL, del.u = NULL, del.l = NULL, add = FALSE)
+visualSelect(ring.data, del = NULL, del.u = NULL, del.l = NULL,
+  add = FALSE)
 }
 \arguments{
-  \item{ring.data}{A matrix or array produced by \code{autoDetect}.}
-  \item{del}{A numeric vector giving the border numbers to be deleted.}
-  \item{del.u}{A numeric vector giving the border numbers to be deleted on the upper path.}
-  \item{del.l}{A numeric vector giving the border numbers to be deleted on the lower path.}
-  \item{add}{A logical value indicating whether to re-add ring borders visually.}
-}
+\item{ring.data}{A matrix or array produced by \code{autoDetect}.}
 
-\value{A matrix or array representing the tree ring image.}
+\item{del}{A numeric vector giving the border numbers to be deleted.}
+
+\item{del.u}{A numeric vector giving the border numbers to be deleted 
+on the upper path.}
+
+\item{del.l}{A numeric vector giving the border numbers to be deleted 
+on the lower path.}
+
+\item{add}{A logical value indicating whether to re-add ring borders.}
+}
+\value{
+A matrix or array representing the tree ring image.
+}
+\description{
+This function can delete existing ring borders 
+or add new borders by clicking on the image.
+}
 \details{
-This function is used to delete existing ring borders, or to add new borders by interactively clicking on the image segments.
+This function is used to delete existing ring borders, or to add new 
+borders by interactively clicking on the image segments.
 
-If the user creates only one path (\code{incline = FALSE}), the argument \code{del} is used to delete ring borders. If the user creates two paths (\code{incline = TRUE}), arguments \code{del.u} and \code{del.l} are used to delete ring borders.
+If the user creates only one path (\code{incline = FALSE}), the argument 
+\code{del} is used to delete ring borders. If the user creates two paths 
+(\code{incline = TRUE}), arguments \code{del.u} and \code{del.l} are used 
+to delete ring borders.
 
-If \code{add = TRUE}, graphical windows where image segments are plotted will be activated sequentially. When a graphical window is activated, the user can add new borders by left-clicking the mouse along the path. This visual selection process can be terminated by clicking the right button and selecting ‘Stop’ from the menu, or from the ‘Stop’ button on the top-left corner of the graphical window (depending on the R version). 
+If \code{add = TRUE}, graphical windows where image segments are plotted 
+will be activated sequentially. When a graphical window is activated, 
+the user can add new borders by left-clicking the mouse along the path. 
+This visual selection process can be terminated by clicking the right 
+button and selecting "Stop" from the menu, or from the "Stop" button on 
+the top-left corner of the graphical window (depending on the R version). 
 
-Once the user terminates the visual selection process, the current graphical window will be closed automatically, and the graphical window of the following segment is activated. When all graphical windows are closed, this function will re-open graphical windows and plot new borders.
+Once the user terminates the visual selection process, the current 
+graphical window will be closed automatically, and the graphical window 
+of the following segment is activated. When all graphical windows are 
+closed, this function will re-open graphical windows and plot new borders.
 
[…]
