+++ man/MtreeRing-package.Rd
@@ -2,9 +2,16 @@
 \alias{MtreeRing-package}
 \alias{MtreeRing}
 \docType{package}
-\title{A Shiny Application for Automatic Measurements of Tree-Ring Widths on Digital Images}
+\title{A Shiny Application for Automatic Measurements of Tree-Ring Widths 
+on Digital Images}
 \description{
-Use morphological image processing and edge detection algorithms to automatically measure tree ring widths on digital images. Users can also manually mark tree rings on species with complex anatomical structures. The arcs of inner-rings and angles of successive inclined ring boundaries are used to correct ring-width series. The package provides a Shiny-based application, allowing R beginners to easily analyze tree ring images and export ring-width series in standard file formats.
+Use morphological image processing and edge detection algorithms to 
+automatically measure tree ring widths on digital images. Users can also 
+manually mark tree rings on species with complex anatomical structures. 
+The arcs of inner-rings and angles of successive inclined ring boundaries 
+are used to correct ring-width series. The package provides a Shiny-based 
+application, allowing R beginners to easily analyze tree ring images and 
+export ring-width series in standard file formats.
 }
 \details{
 
