+++ R/nearPith.R
@@ -1,40 +1,65 @@
 #' @title Calibrate ring-width series
 #' @export
 #' @importFrom stats coef lm
-#' @description This function can calibrate the ring-width series using arcs of inner rings.
+#' @description This function can calibrate the ring-width series 
+#' using arcs of inner rings.
 #' @author Jingning Shi
 #' @param ring.data A magick image object produced by \code{imgInput}.
-#' @param inner.arc A logical value indicating whether to calibrate the ring-width series using the arcs of inner rings. See details below.
-#' @param last.yr \code{NULL} or an integer giving the year of formation of the left-most ring. If \code{NULL}, borders numbers (starting from 1) ares used instead of the year.
+#' @param inner.arc A logical value indicating whether to calibrate the 
+#' ring-width series using the arcs of inner rings. See details below.
+#' @param last.yr \code{NULL} or an integer giving the year of formation 
+#' of the left-most ring. If \code{NULL}, borders numbers (starting from 1) 
+#' are used instead of the year.
 #' @param color Color for labels.
-#' @param border.type Symbol for ring borders. See \code{pch} in \code{\link{points}} for possible values and their shapes.
+#' @param border.type Symbol for ring borders. See \code{pch} in 
+#' \code{\link{points}} for possible values and their shapes.
 #' @param label.cex The magnification to be used for years and ring numbers.
 #' @return A data frame of the calibrated ring-width series.
 #' @details 
-#' This function allows the user to create a path, and visually identify ring borders by clicking on the graphical window.  
+#' This function allows the user to create a path, and manually mark 
+#' ring borders by clicking on the graphical window.  
 #' \itemize{
 #' \item
-#' If \code{inner.arc = TRUE}, the ring-width series is calibrated using arcs of inner rings (Duncan, 1989). 
-#' First, the user can click the left mouse button to add a horizontal path. The path should traverse an appropriate arc (read the reference below for more details). Then, the user can add three points to the selected arc by left-clicking. The first point should be placed on the left endpoint of the arc, and the second point is placed on the right endpoint. 
+#' If \code{inner.arc = TRUE}, the ring-width series is calibrated using arcs 
+#' of inner rings (Duncan, 1989).
+#'  
+#' First, the user can click the left mouse button to add a horizontal path.
+#' The path should traverse an appropriate arc (read the reference below for 
+#' more details). Then, the user can add three points to the selected arc by
+#' left-clicking. The first point should be placed on the left endpoint of 
+#' the arc, and the second point is placed on the right endpoint. 
 #' 
-#' After two endpoints are added, a vertical dashed line will be plotted automatically according to the (x,y) positions of endpoints you just added. The third points should be placed on the intersection of the vertical dashed line and the selected arc. 
+#' After two endpoints are added, a vertical dashed line will be plotted 
+#' automatically according to the (x,y) positions of endpoints you just added. 
+#' The third points should be placed on the intersection of the vertical 
+#' dashed line and the selected arc. 
 #' 
-#' Finally, the user is prompted to visually mark ring borders along the path. The termination of visual selection is similar to \code{visualSelect}. Note that the left endpoint of the arc will be considered as the last ring border without the need to mark it.
+#' Finally, the user is prompted to visually mark ring borders along the path.
+#' The termination of visual selection is similar to \code{visualSelect}.
+#' Note that the left endpoint of the arc will be considered as the last 
[…]
