+++ tests/testthat/test-launchMtRApp.R
@@ -1,7 +1,7 @@
 context("launchMtRApp")
 
 test_that("the function run a app", {
-  
+ 
   a <- launchMtRApp()
   a
   expect_is(a, "shiny.appobj")
