+++ NAMESPACE
@@ -6,12 +6,12 @@ export(imgInput)
 export(launchMtRApp)
 export(nearPith)
 export(visualSelect)
-import(dplyr)
-import(shiny)
-import(shinyWidgets)
-import(shinydashboard)
 importFrom(bmp,read.bmp)
 importFrom(dplR,write.rwl)
+importFrom(dplyr,filter)
+importFrom(dplyr,group_by)
+importFrom(dplyr,select)
+importFrom(dplyr,summarize)
 importFrom(grDevices,as.raster)
 importFrom(grDevices,dev.cur)
 importFrom(grDevices,dev.list)
@@ -50,6 +50,56 @@ importFrom(magrittr,"%>%")
 importFrom(measuRing,graySmoothed)
 importFrom(measuRing,linearDetect)
 importFrom(png,readPNG)
+importFrom(shiny,actionButton)
+importFrom(shiny,br)
+importFrom(shiny,brushOpts)
+importFrom(shiny,checkboxInput)
+importFrom(shiny,column)
+importFrom(shiny,conditionalPanel)
+importFrom(shiny,dblclickOpts)
+importFrom(shiny,div)
+importFrom(shiny,downloadButton)
+importFrom(shiny,downloadHandler)
+importFrom(shiny,fileInput)
+importFrom(shiny,fluidRow)
+importFrom(shiny,helpText)
+importFrom(shiny,hr)
+importFrom(shiny,icon)
+importFrom(shiny,numericInput)
+importFrom(shiny,observeEvent)
+importFrom(shiny,plotOutput)
+importFrom(shiny,reactiveValues)
+importFrom(shiny,renderPlot)
+importFrom(shiny,renderTable)
+importFrom(shiny,selectInput)
+importFrom(shiny,shinyApp)
+importFrom(shiny,sliderInput)
+importFrom(shiny,tabPanel)
+importFrom(shiny,tableOutput)
+importFrom(shiny,textInput)
+importFrom(shiny,updateActionButton)
+importFrom(shiny,updateCheckboxInput)
+importFrom(shiny,updateTextInput)
+importFrom(shinyWidgets,colorSelectorInput)
+importFrom(shinyWidgets,prettyCheckbox)
+importFrom(shinyWidgets,prettyRadioButtons)
+importFrom(shinyWidgets,prettySwitch)
+importFrom(shinyWidgets,radioGroupButtons)
+importFrom(shinyWidgets,sendSweetAlert)
+importFrom(shinyWidgets,updatePrettyCheckbox)
+importFrom(shinyWidgets,updatePrettyRadioButtons)
+importFrom(shinyWidgets,updatePrettySwitch)
+importFrom(shinyWidgets,useSweetAlert)
+importFrom(shinydashboard,box)
+importFrom(shinydashboard,dashboardBody)
+importFrom(shinydashboard,dashboardHeader)
+importFrom(shinydashboard,dashboardPage)
+importFrom(shinydashboard,dashboardSidebar)
+importFrom(shinydashboard,menuItem)
+importFrom(shinydashboard,sidebarMenu)
+importFrom(shinydashboard,tabBox)
+importFrom(shinydashboard,tabItem)
+importFrom(shinydashboard,tabItems)
 importFrom(spatstat,connected)
 importFrom(spatstat,im)
 importFrom(stats,coef)
