### 2019-03-19 #287 @JingningShi
Thanks for checking out the package @annakrystalli . English isn't my first language, so please excuse any mistakes. I made the following changes based on your comments:
1. Generate the ```NAMESPACE``` and ```man/``` files using ```devtools::document()```.
2. Replace the images which throw up warnings with new image files.
  Sometimes the test will issue a warning ```libpng warning: iCCP: known incorrect sRGB profile```. This happens because of changes in libpng version 1.6+, and will not affect the R function.
3.  Add a blank line at the end of  ```DESCRIPTION``` file and ```R code files``` .
4. Create  a ```buildignore``` file using the command ```usethis::use_build_ignore(".travis.yml")```.
5. Replace the "Depends" in DESCRIPTION with “Imports”.
6. Remove sapply() as suggested by ```goodpractice```.
7. Import only the specific functions we need, rather than the whole package.

Regarding test coverage you mentioned, the reason is that some functions in our package use  ```loactor()``` to record mouse position. If running tests which call ```locator()```, it produces error messages, such as ```nearPith()```.  Another reason is that ```test_that()``` can't test the Shiny app.  The code file of the app (createServer.R) has 1571 lines (the package has 2868 lines in all) so the test coverage is less than 75%.

Thanks again for looking it over and please let me know if you have any other questions.
