+++ .travis.yml
@@ -5,10 +5,9 @@ r:
   - release
   - devel 
 before_install:
+  - sudo apt-get update
   - sudo apt-get install -y libmagick++-dev
   - sudo add-apt-repository -y ppa:opencpu/imagemagick
-  - sudo apt-get update
-  - sudo apt-get install -y libmagick++-dev             
 r_packages:
   - covr
 addons:
@@ -31,4 +30,4 @@ addons:
 cache: packages
 
 after_success:
-  - Rscript -e 'covr::codecov(function_exclusions='createServer')'
+  - Rscript -e 'covr::codecov()'
