+++ .travis.yml
@@ -4,10 +4,7 @@ language: r
 r:
   - release
   - devel 
-before_install:
-  - sudo apt-get update
-  - sudo apt-get install -y libmagick++-dev
-  - sudo add-apt-repository -y ppa:opencpu/imagemagick
+
 r_packages:
   - covr
 addons:
