+++ .travis.yml
@@ -4,7 +4,10 @@ language: r
 r:
   - release
   - devel 
-
+before_install:
+  - sudo apt-get update
+  - sudo add-apt-repository -y ppa:opencpu/imagemagick
+  - sudo apt-get install -y libmagick++-dev
 r_packages:
   - covr
 addons:
