### 2019-03-19 #287 @annakrystalli
Hi @JingningShi ,

Thanks for your speedy and full response!

Regarding English as a second language, don't worry at all. If there is anything I write that doesn't make sense let me know and I will try to explain in a different way. I will do the same. ☺️

I've re-run all the checks & tests and most of the issues have been solved! 🎉

Regarding testing coverage, I have two suggestions to increase it:
1.  Have you seen [shinytest](https://rstudio.github.io/shinytest/articles/shinytest.html) for testing shiny apps? This should allow to test the behaviour of the shinyapp.
2. You could use [`mockery`](https://github.com/jfiksel/mockery) package to set the output of `locator()` during testing so you can check chunks which call locator().

Finally, `goodpractice::gp()` still returns a warning about the [cyclomatic complexity](https://www.softwaretestinghelp.com/cyclomatic-complexity/) of functions: `createServer (254), autoDetect (52).` Do you think the internals of these functions could be simpler?
