+++ inst/mtr_app/app.R
@@ -1,0 +1,2156 @@
+library(shiny)
+library(shinyWidgets)
+library(shinydashboard)
+library(MtreeRing)
+library(testthat)
+library(magrittr)
+library(png)
+library(jpeg)
+library(tiff)
+library(bmp)
+library(magick)
+library(imager)
+library(dplR)
+library(spatstat)
+library(measuRing)
+library(dplyr)
+
+
+createUI <- function()
+{
+  shiny.title <- dashboardHeader(title = 'MtreeRing')
+  shiny.sider <- dashboardSidebar(
+    sidebarMenu(
+      menuItem('Image Loading',tabName = 'input_pre', 
+        icon = icon('folder-open', lib = 'font-awesome')),
+      menuItem('Measurement',tabName = 'mea_arg', 
+        icon = icon('gear', lib = 'font-awesome'))
+    )
+  )
+  page1 <- fluidRow(
+    box(
+      title = div(style = 'color:#FFFFFF;font-size:80%; 
+        font-weight: bolder', 'Image Preview'),
+      width = 12, status = 'primary', solidHeader = T, collapsible = T,
+      plotOutput('pre.img',
+        brush = brushOpts(
+          id = "plot1_brush",
+          opacity = 0.25,
+          resetOnNew = TRUE)
+      )
+      ), 
+    box(
+      title = div(style = 'color:#FFFFFF;font-size:80%;
+        font-weight: bolder', 'Image Upload'),
+      width = 4, status = 'primary', solidHeader = T, collapsible = T,
+      conditionalPanel(
+        condition = '!input.inmethod',
+        fileInput('select.file', 'Choose an image file',
+          buttonLabel = 'Browse...', width = '80%')
+      ),
+      prettySwitch(inputId = "magick.switch", label = "Magick ON",
+        value = TRUE, fill = TRUE, status = "success"),
+      helpText('Image upload is limited to 150 MB per file. Supported',
+        ' formats include png, jpg, tif and bmp.',
+        style = 'color:#000000;font-size:90%'),
+      prettyCheckbox(
+        inputId = "inmethod", 
+        label = div(style = 'color:#000000;font-weight: bolder;', 'Image Path'), 
+        shape = "curve", value = F, status = "success"),
+      conditionalPanel(
+        condition = 'input.inmethod',
+        textInput('enter.path', 'Enter file path', ''),
+        helpText('For example: C:/Users/shiny/img01.png',
+          style = 'color:#000000;font-size:90%'),
+        hr()
+      ),
+      actionButton(
+        'buttoninputimage', 'Load ',
+        class = "btn btn-primary btn-md",
+        icon = icon('upload',  "fa-1x"),
+        style = 'color:#FFFFFF;text-align:center;
+        font-weight: bolder;font-size:110%;'),
+      useSweetAlert()
+      ),
+    box(
+      title = div(style = 'color:#FFFFFF;font-size:80%;
+        font-weight: bolder', 'Image Rotation'),
+      width = 3, status = 'primary', solidHeader = T, collapsible = T,
+      prettyRadioButtons(inputId = "rotatede", label = "",
+        choices = c("0 degrees" = "rotate0",
+          "90 degrees" = "rotate90",
+          "180 degrees" = "rotate180",
+          "270 degrees" = "rotate270"),
+        shape = "curve", status = "success",
+        fill = TRUE, inline = TRUE),
+      helpText("Rotation angle in degrees. Note that the bark ",
+        "side should be placed at the left side of the ",
+        "graphical window and the pith side at the right.",
+        style = 'color:#000000;font-size:90%;text-align:justify;'),
+      actionButton(
+        'buttonrotate', 'Rotate',
[…]
