+++ inst/mtr_app/tests/upload_image.R
@@ -1,7 +1,8 @@
 app <- ShinyDriver$new("../")
 app$snapshotInit("upload_image")
+
 app$setInputs(sidebarCollapsed = FALSE)
-app$setInputs(inmethod = TRUE)
-app$setInputs(`enter.path` = system.file('001.png', package = 'MtreeRing'))
+# Input 'plot1_brush' was set, but doesn't have an input binding.
+app$uploadFile(selectfile = system.file('001.png', package = 'MtreeRing'))
 app$setInputs(buttoninputimage = "click")
 app$snapshot()
