+++ R/createUI.R
@@ -1,560 +1,0 @@
-#' @importFrom shiny actionButton checkboxInput div fileInput conditionalPanel
-#' downloadButton downloadHandler helpText icon numericInput observeEvent
-#' plotOutput reactiveValues renderPlot renderTable selectInput shinyApp
-#' sliderInput tableOutput tabPanel textInput updateActionButton brushOpts
-#' dblclickOpts column
-#' updateCheckboxInput updateTextInput hr br fluidRow
-#' @importFrom shinydashboard dashboardHeader dashboardSidebar dashboardBody
-#' dashboardPage sidebarMenu tabBox tabItem tabItems menuItem box
-#' @importFrom  shinyWidgets prettySwitch prettyCheckbox prettyRadioButtons
-#' radioGroupButtons sendSweetAlert updatePrettySwitch updatePrettyCheckbox
-#' updatePrettyRadioButtons useSweetAlert colorSelectorInput
-
-createUI <- function()
-{
-shiny.title <- dashboardHeader(title = 'MtreeRing')
-shiny.sider <- dashboardSidebar(
-  sidebarMenu(
-    menuItem('Image Loading',tabName = 'input_pre', 
-             icon = icon('folder-open', lib = 'font-awesome')),
-    menuItem('Measurement',tabName = 'mea_arg', 
-             icon = icon('gear', lib = 'font-awesome'))
-  )
-)
-page1 <- fluidRow(
-  box(
-    title = div(style = 'color:#FFFFFF;font-size:80%; 
-                font-weight: bolder', 'Image Preview'),
-    width = 12, status = 'primary', solidHeader = T, collapsible = T,
-    plotOutput('pre.img',
-               brush = brushOpts(
-                 id = "plot1_brush",
-                 opacity = 0.25,
-                 resetOnNew = TRUE)
-    )
-  ), 
-  box(
-    title = div(style = 'color:#FFFFFF;font-size:80%;
-                font-weight: bolder', 'Image Upload'),
-    width = 4, status = 'primary', solidHeader = T, collapsible = T,
-    conditionalPanel(
-      condition = '!input.inmethod',
-      fileInput('select.file', 'Choose an image file',
-                buttonLabel = 'Browse...', width = '80%')
-    ),
-    prettySwitch(inputId = "magick.switch", label = "Magick ON",
-                 value = TRUE, fill = TRUE, status = "success"),
-    helpText('Image upload is limited to 150 MB per file. Supported',
-             ' formats include png, jpg, tif and bmp.',
-             style = 'color:#000000;font-size:90%'),
-    prettyCheckbox(
-      inputId = "inmethod", 
-      label = div(style = 'color:#000000;font-weight: bolder;', 'Image Path'), 
-      shape = "curve", value = F, status = "success"),
-    conditionalPanel(
-      condition = 'input.inmethod',
-      textInput('enter.path', 'Enter file path', ''),
-      helpText('For example: C:/Users/shiny/img01.png',
-               style = 'color:#000000;font-size:90%'),
-      hr()
-    ),
-    actionButton(
-      'buttoninputimage', 'Load ',
-      class = "btn btn-primary btn-md",
-      icon = icon('upload',  "fa-1x"),
-      style = 'color:#FFFFFF;text-align:center;
-              font-weight: bolder;font-size:110%;'),
-    useSweetAlert()
-  ),
-  box(
-    title = div(style = 'color:#FFFFFF;font-size:80%;
-                font-weight: bolder', 'Image Rotation'),
-    width = 3, status = 'primary', solidHeader = T, collapsible = T,
-    prettyRadioButtons(inputId = "rotatede", label = "",
-                       choices = c("0 degrees" = "rotate0",
-                                   "90 degrees" = "rotate90",
-                                   "180 degrees" = "rotate180",
[…]
