+++ NAMESPACE
@@ -73,6 +73,7 @@ importFrom(shiny,renderPlot)
 importFrom(shiny,renderTable)
 importFrom(shiny,selectInput)
 importFrom(shiny,shinyApp)
+importFrom(shiny,shinyAppDir)
 importFrom(shiny,sliderInput)
 importFrom(shiny,tabPanel)
 importFrom(shiny,tableOutput)
