+++ DESCRIPTION
@@ -5,7 +5,7 @@ Version: 1.2
 Authors@R: c(person("Jingning", "Shi", role = c("aut", "cre"), email = "snow940220@bjfu.edu.cn"), person("Wei", "Xiang", role = "aut"))
 Author: Jingning Shi [aut, cre], Wei Xiang [aut]
 Maintainer: Jingning Shi <snow940220@bjfu.edu.cn>
-Suggests: testthat, knitr, Shinytest
+Suggests: testthat, knitr, shinytest
 Imports: magrittr (>= 1.5), png, jpeg, tiff, bmp, magick, imager, dplR, spatstat, measuRing, shiny, dplyr, shinydashboard, shinyWidgets
 Description: Use morphological image processing and edge detection algorithms to automatically measure tree ring widths on digital images. Users can also manually mark tree rings on species with complex anatomical structures. The arcs of inner-rings and angles of successive inclined ring boundaries are used to correct ring-width series. The package provides a Shiny-based application, allowing R beginners to easily analyze tree ring images and export ring-width series in standard file formats.
 License: GPL-3
