+++ inst/doc/introduction.Rmd
@@ -1,75 +1,0 @@
----
-title: "MtreeRing: Measuring Tree Ring Width in R"
-author: "Jingning Shi"
-date: "2019-03-01"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{downloading}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-<!--require(knitr);require(markdown);knit("introduction.Rmd")-->
-
-`MtreeRing` is a tool for automatically measuring tree-ring width using image processing techniques.
-
-## Installation
-
-Install the stable version from CRAN
-
-```{r eval=FALSE}
-install.packages("MtreeRing")
-```
-
-or the development version from GitHub
-
-```{r eval=FALSE}
-# install.packages("devtools")
-devtools::install_github("JingningShi/MtreeRing")
-```
-
-## Read and plot a tree ring image
-
-```{r}
-library(MtreeRing)
-## Read and plot a tree ring image
-img.name <- system.file("001.png", package = "MtreeRing")
-t1 <- imgInput(img = img.name, dpi = 1200)
-```
-
-`imgInput` supports commonly used image formats and returns a magick image object.
-
-
-## Detect ring borders 
-
-After plotting the image, the automatic detection of ring borders can be performed using three alternative methods: (1) watershed algorithm; (2) Canny edge detector; (3) a linear detection algorithm from R package [measuRing](https://CRAN.R-project.org/package=measuRing).
-
-If `auto.path = TRUE`, the function will create a path at the center of the image. In the current version, the path is a horizontal dashed line. Detected ring borders are placed along the path, and are tagged with years and border numbers.
-
-```{r}
-## Split a long core sample into 2 pieces to
-## get better display performance and use the
-## watershed algorithm to detect ring borders:
-t2 <- autoDetect(ring.data = t1, auto.path = TRUE, seg = 2, method = 'watershed')
-```
-
-
-For increment cores with inclined rings, you can create two paths by setting `incline = TRUE`. The distance between adjacent rings and orientation of ring borders are used to correct the ring-width series.
-
-```{r}
-img.name <- system.file("incline.png", package = "MtreeRing")
-t1 <- imgInput(img = img.name, dpi = 1200)
-t2 <- autoDetect(ring.data = t1, incline = TRUE, method = 'watershed')
-```
-
-When ring borders are added to the image, you can generate a ring-width series in dataframe format. Use `write.rwl` (from R package `dplR`) to export this dataframe in rwl format.
-
-```{r}
-rw.df <- calcRingWidth(ring.data = t2, seriesID = "940220")
-```
-
-If you are not familiar with R and its command line interface, use the shiny-based app. A workflow for the Shiny app can be found in the README file of this package. Here is a link: https://github.com/JingningShi/MtreeRing.
-
-```{r eval=FALSE}
-launchMtRApp()
-```
