### 2019-03-20 da6296c
Delete run_tests.R
