+++ inst/mtr_app/run_tests.R
@@ -1,21 +1,0 @@
-library(testthat)
-library(shinytest)
-library(shiny)
-library(shinyWidgets)
-library(shinydashboard)
-library(MtreeRing)
-library(magrittr)
-library(png)
-library(jpeg)
-library(tiff)
-library(bmp)
-library(magick)
-library(imager)
-library(dplR)
-library(spatstat)
-library(measuRing)
-library(dplyr)
-
-test_that("Application works", {
-  expect_pass(testApp(".", compareImages = FALSE))
-})
