+++ tests/testthat/test-launchMtRApp.R
@@ -3,7 +3,6 @@ context("launchMtRApp")
 test_that("the function run a app", {
  
   a <- launchMtRApp()
-  a
   expect_is(a, "shiny.appobj")
   expect_type(a, "list")
   

[…]
