+++ .travis.yml
@@ -11,6 +11,7 @@ before_install:
   - sudo apt-get install -y libmagick++-dev         
 r_packages:
   - covr
+  - mockery
 addons:
   apt:
     packages:
