+++ R/nearPith.R
@@ -82,54 +82,9 @@ nearPith <- function(ring.data, inner.arc = TRUE, last.yr = NULL,
   dp <- x.dpi / 25.4 
   dev.set(device.number)
   dimt <- attributes(ring.data)$dimt
-  text.line <- 4
-  step.number <- 1
-  if (inner.arc) {
-    mtext(paste0('Step ', step.number, ': Click the left ',
-                 'mouse button to add a horizontal path.'), 
-          side = 1, line = text.line, adj = 0)
-    step1 <- locator(n = 1, type = "n")
-    py <- round(step1$y) 
-    abline(h = py, lty = 2, lwd = 2, col = color)
-    text.line <- 2 + text.line
-    mtext(paste0("Step ", step.number, " is already done."), side = 1, 
-          line = text.line, adj = 0, col = "blue")
-    text.line <- 2 + text.line
-    step.number <- 1 + step.number
-    mtext(paste0('Step ', step.number, ' : Click the left mouse button',
-                 ' to add three points on the arc you select.'), 
-          side = 1, line = text.line, adj = 0)
-    arc.a <- locator(1, type = "n")
-    points(arc.a$x, py, pch = border.type, col = color, cex = label.cex)
-    arc.b <- locator(1, type = "n")
-    points(arc.b$x, py, pch = border.type, col = color, cex = label.cex)
-    ab.x <- (arc.a$x + arc.b$x)/2
-    abline(v = ab.x, lwd = 1, lty = 2, col = color)
-    arc.c <- locator(1, type = "n")
-    points(ab.x, arc.c$y, pch = border.type, col = color, cex = label.cex)
-    text.line <- 2 + text.line
-    mtext(paste0("Step ", step.number, " is already done "), side = 1, 
-          line = text.line, adj = 0, col = "blue")
-  }  
-  if (!inner.arc) {
-    mtext(paste0('Step ', step.number, ' :Click the left mouse button to add ', 
-                 'two points. A straight line passing through ',
-                 'these two points will be added to the image.'), 
-          side = 1, line = text.line, adj = 0)
-    p1 <- locator(1, type = 'p', pch = 19)
-    points(p1$x, p1$y, pch = border.type, col = color, cex = label.cex)
-    p2 <- locator(1, type = 'p', pch = 19)
-    points(p2$x, p2$y, pch = border.type, col = color, cex = label.cex)
-    path.x <- c(p1$x, p2$x)
-    path.y <- c(p1$y, p2$y)
-    path.xy <- lm(path.y ~ path.x + 1)
-    a <- coef(path.xy)[1]
-    b <- coef(path.xy)[2]
-    abline(a = a, b = b, lwd = 2, lty = 2, col = color)
-    text.line <- 2 + text.line
-    mtext(paste0('Step ', step.number, ' is already done.'), side = 1, 
-          line = text.line, adj = 0, col = 'blue')
-  }  
+  pos <- add_path(inner.arc, border.type, color, label.cex)
+  text.line <- pos$t
+  step.number <- pos$s
   text.line <- 2 + text.line
   mtext(text = 'Please mark ring boundaries along the user-defined path', 
         side = 1, line = text.line, adj = 0)
@@ -138,6 +93,10 @@ nearPith <- function(ring.data, inner.arc = TRUE, last.yr = NULL,
   bor.x <- sort(bor.xy$x)
   bor.y <- bor.xy$y[order.x]
   if (inner.arc) {
+    arc.a <- pos$arc$a
+    arc.b <- pos$arc$b
+    arc.c <- pos$arc$c
+    py <- pos$p
[…]
