+++ tests/testthat/test-nearPith.R
@@ -1,0 +1,132 @@
+context("nearPith")
+##test1
+test_that("nearPith returns a dataframe", {
+  
+  library(mockery)
+  path1 <- system.file("arc.png", package = "MtreeRing")
+  ## Read and plot the image:
+  img1 <- imgInput(img = path1, dpi = 1200)
+
+  mock1 <- mock(list(x = c(36, 110, 168.8, 249.3, 354.2), 
+                     y = c(264, 264, 264, 264, 264)), cycle = T)
+  stub(nearPith, 'locator', mock1)
+  
+  a <- list(x = 444.7, y = 264.46)
+  b <- list(x = 643, y = 264.46)
+  c <- list(x = 543.8, y = 300.48)
+  arc <- list(a = a, b = b, c = c)
+
+  mock2 <- mock(list(arc = arc, p = 264, t = 10, s = 2), cycle = T)
+  stub(nearPith, 'add_path', mock2)
+  t2 <- nearPith(img1, T)
+  
+  dev.off(attributes(img1)$dn)
+  expect_is(img1, "magick-image")
+  expect_is(t2, "data.frame")
+})
+##test2
+test_that("nearPith returns a dataframe", {
+  
+  library(mockery)
+  path1 <- system.file("arc.png", package = "MtreeRing")
+  ## Read and plot the image:
+  img1 <- imgInput(img = path1, dpi = 1200)
+  
+  mock1 <- mock(list(x = c(36, 110, 168.8, 249.3, 354.2), 
+    y = c(264, 264, 264, 264, 264)), cycle = T)
+  stub(nearPith, 'locator', mock1)
+  
+  a <- list(x = 444.7, y = 264.46)
+  b <- list(x = 643, y = 264.46)
+  c <- list(x = 543.8, y = 300.48)
+  arc <- list(a = a, b = b, c = c)
+  
+  mock2 <- mock(list(arc = arc, p = 264, t = 10, s = 2), cycle = T)
+  stub(nearPith, 'add_path', mock2)
+  t2 <- nearPith(img1, T, last.yr = 2015)
+  
+  dev.off(attributes(img1)$dn)
+  expect_is(img1, "magick-image")
+  expect_is(t2, "data.frame")
+})
+##test3
+test_that("nearPith returns a dataframe", {
+  
+  library(mockery)
+  path1 <- system.file("arc.png", package = "MtreeRing")
+  ## Read and plot the image:
+  img1 <- imgInput(img = path1, dpi = 1200)
+  
+  mock1 <- mock(list(t = 6, s = 1), cycle = T)
+  stub(nearPith, 'add_path', mock1)
+  
+  mock2 <- mock(
+   list(x = c(33.39, 101.03, 151.16, 223.58, 305.55, 369.21, 431.28, 471.86), 
+        y = c(250.25, 233.19, 223.71, 202.86, 182.00, 167.79, 148.83, 140.30)), 
+   cycle = T)
+  stub(nearPith, 'locator', mock2)
+  t2 <- nearPith(img1, F, last.yr = 2015)
+  dev.off(attributes(img1)$dn)
+  expect_is(img1, "magick-image")
+  expect_is(t2, "data.frame")
+})
+##test4
+test_that("nearPith returns a dataframe", {
+  
+  library(mockery)
+  path1 <- system.file("arc.png", package = "MtreeRing")
+  ## Read and plot the image:
+  img1 <- imgInput(img = path1, dpi = 1200)
+  
+  mock1 <- mock(list(t = 6, s = 1), cycle = T)
+  stub(nearPith, 'add_path', mock1)
+  
+  mock2 <- mock(
[…]
