+++ .travis.yml
@@ -1,12 +1,24 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
 language: r
+os:
+  - linux
+  - osx
 r:
   - release
   - devel 
 before_install:
-  - sudo apt-get install -y libmagick++-dev
-  - sudo add-apt-repository -y ppa:opencpu/imagemagick
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then sudo apt-get -qq update; fi
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then sudo apt-get install -y libmagick++-dev; fi
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then sudo apt-get install libtiff4-dev; fi
+  - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then sudo apt-get install libfftw3-dev; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew reinstall imagemagick@6 --with-fontconfig --with-librsvg; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew link --force imagemagick@6; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew install libtiff; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew cask uninstall oclint; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew install gcc || brew link --overwrite gcc; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew install fftw; fi
   
 r_packages:
   - covr
