+++ .travis.yml
@@ -7,7 +7,9 @@ r:
   - release
   - devel 
 before_install:
-  - sudo apt-get -qq update
+  - sudo apt-get install -y libmagick++-dev
+  - sudo add-apt-repository -y ppa:opencpu/imagemagick
+  - sudo apt-get update
   - sudo apt-get install -y libmagick++-dev
   
 r_packages:
