+++ inst/mtr_app/app.R
@@ -56,7 +56,7 @@ createUI <- function()
         style = 'color:#000000;font-size:90%'),
       prettyCheckbox(
         inputId = "inmethod", 
-        label = div(style = 'color:#000000;font-weight: bolder;', 'Image Path'), 
+        label = div(style = 'color:#000000;font-weight: bolder;','Image Path'), 
         shape = "curve", value = F, status = "success"),
       conditionalPanel(
         condition = 'input.inmethod',
@@ -147,8 +147,8 @@ createUI <- function()
       width = 4, status = 'primary', solidHeader = T, collapsible = T,
       textInput('tuid', 'Series ID', '', width = '75%'),
       textInput('dpi', 'DPI', '', '75%'),
-      textInput('sample.yr', 'Sampling year', '', '75%'),
-      textInput('m.line', 'Y-coordinate of path', '', '75%'),
+      textInput('sample_yr', 'Sampling year', '', '75%'),
+      textInput('m_line', 'Y-coordinate of path', '', '75%'),
       prettyCheckbox(
         inputId = "incline", 
         label = div(
@@ -675,7 +675,7 @@ createServer <- function(input, output, session)
     border.col <- f.sort(border.col, dp)
     return(border.col)
   }
-  plot.marker <- function(py, incline, dp, sample.yr, h.dis, l.w, 
+  plot.marker <- function(py, incline, dp, sample_yr, h.dis, l.w, 
     bor.color, lab.color, pch, label.cex, 
     df.loc, plot.year, img.name)
   {
@@ -706,7 +706,7 @@ createServer <- function(input, output, session)
           points(bx[up], by.up, col = bor.color, type = "p", 
             pch = pch, cex = label.cex * 0.75)
           if (plot.year) {
-            year.u <- c(sample.yr:(sample.yr - lenup + 1))
+            year.u <- c(sample_yr:(sample_yr - lenup + 1))
             text(bx[up], by.up, year.u, adj = c(1.5, 0.5), 
               srt = 90, col = lab.color, cex = label.cex)
             border.num <- 1:lenup
@@ -721,7 +721,7 @@ createServer <- function(input, output, session)
           points(bx[lower], by.lower, col = bor.color, type = "p", 
             pch = pch, cex = label.cex * 0.75)
           if (plot.year) {
-            year.l <- c(sample.yr:(sample.yr - lenlo + 1))
+            year.l <- c(sample_yr:(sample_yr - lenlo + 1))
             text(bx[lower], by.lower, year.l, adj = c(1.5, 0.5), 
               srt = 90, col = lab.color, cex = label.cex)
             border.num <- 1:lenlo
@@ -736,7 +736,7 @@ createServer <- function(input, output, session)
           points(bx, by, col = bor.color, type = "p", 
             pch = pch, cex = label.cex * 0.75)
           if (plot.year) {
-            year.u <- c(sample.yr:(sample.yr - length(by) + 1))
+            year.u <- c(sample_yr:(sample_yr - length(by) + 1))
             text(bx, by, year.u, adj = c(1.5, 0.5), 
               srt = 90, col = lab.color, cex = label.cex)
             border.num <- 1:lenbx
@@ -747,7 +747,7 @@ createServer <- function(input, output, session)
       }
     }
   }
-  f.rw <- function(outfile, sample.yr, incline, py, dpi, h.dis) {
+  f.rw <- function(outfile, sample_yr, incline, py, dpi, h.dis) {
     df.loc <- outfile
     bx <- df.loc$x[-c(1:2)]
     where.bx <- df.loc$z[-c(1:2)]
[…]
