+++ .travis.yml
@@ -6,11 +6,14 @@ os:
 r:
   - release
   - devel 
+
+  
 before_install:
+  - sudo apt-get -qq update
   - sudo apt-get install -y libmagick++-dev
   - sudo add-apt-repository -y ppa:opencpu/imagemagick
-  - sudo apt-get update
   - sudo apt-get install -y libmagick++-dev
+
   
 r_packages:
   - covr
