+++ .travis.yml
@@ -9,7 +9,6 @@ r:
 
   
 before_install:
-
   - sudo apt-get install -y libmagick++-dev
   - sudo add-apt-repository -y ppa:opencpu/imagemagick
   - sudo apt-get install -y libmagick++-dev
