+++ tests/testthat/test-app.R
@@ -2,22 +2,7 @@ context('test app')
 
 test_that("app works", {
   library(shinytest)
-  library(shiny)
-  library(shinyWidgets)
-  library(shinydashboard)
-  library(MtreeRing)
-  library(testthat)
-  library(magrittr)
-  library(png)
-  library(jpeg)
-  library(tiff)
-  library(bmp)
-  library(magick)
-  library(imager)
-  library(dplR)
-  library(spatstat)
-  library(measuRing)
-  library(dplyr)
+  skip_on_cran()
   app_dir <- system.file('mtr_app/', package = 'MtreeRing')
   expect_pass(testApp(app_dir, compareImages = FALSE))
 })
