+++ .travis.yml
@@ -8,10 +8,12 @@ r:
   - devel 
 
   
+
 before_install:
   - sudo apt-get install -y libmagick++-dev
   - sudo add-apt-repository -y ppa:opencpu/imagemagick
-  - sudo apt-get install -y libmagick++-dev
+  - sudo apt-get update
+  - sudo apt-get install -y libmagick++-dev  
 
   
 r_packages:
