+++ .travis.yml
@@ -7,18 +7,17 @@ r:
   - release
   - devel 
 
-  
-
 before_install:
   - sudo apt-get install -y libmagick++-dev
   - sudo add-apt-repository -y ppa:opencpu/imagemagick
   - sudo apt-get update
   - sudo apt-get install -y libmagick++-dev  
 
-  
 r_packages:
   - covr
   - mockery
+  - shinytest
+  - testthat
 addons:
   apt:
     packages:
