### 2019-03-21 479d2bf
Delete crop.R
