### 2019-03-21 #287 @JingningShi
Thank you for the suggestions of using ```shinytest``` and ```mockery```. You did me a great favor.
I made the following changes:
1. Use ```mock()``` and ```stub()```  to mock mouse cursor, and the test coverage reaches 97%.
2. Create tests for the app using ```recordTest()``` and ```testApp()```. I save these test scripts and results at a subdirectory of ```inst/``` as suggested by [this article](https://rstudio.github.io/shinytest/articles/package.html).
3. Rewrite the functions you mentioned to reduce cyclomatic complexity. ```goodpractice::gp()``` will not show a warning about the cyclomatic complexity now.

Thanks again for the review. I have learned many things throughout your suggestions. It was extremely helpful.
