### 2019-03-21 8d2c473
Update test-autoDetect.R
