### 2019-03-21 #287 @annakrystalli
Hi @JingningShi,

So glad you found the suggestions useful! And well done on getting coverage up to 97%!! 😃💯
That completes the editor checks. I will start looking for reviewers now.

Could you please add the rOpenSci under review badge to your README?

I'll be back in touch when I have found reviewers 👍
