+++ inst/mtr_app/tests/run_detection2.R
@@ -1,5 +1,7 @@
 app <- ShinyDriver$new("../")
-app$snapshotInit("run_detection")
+app$snapshotInit("run_detection2")
+
+## create two paths in this test
 
 app$setInputs(sidebarCollapsed = FALSE)
 # Input 'plot2_brush' was set, but doesn't have an input binding.
@@ -11,8 +13,9 @@ app$setInputs(dpi = "1200")
 app$setInputs(sample_yr = "2015")
 app$setInputs(m_line = "90")
 app$setInputs(method = 'watershed')
+app$setInputs(incline = TRUE)
 app$setInputs(buttoncreatpath2 = "click")
 # Input 'plot2_brush' was set, but doesn't have an input binding.
-app$setInputs(plot2_brush = list(xmin = 10, xmax = 2333, ymin = 60, ymax = 120),allowInputNoBinding_ = TRUE)
+app$setInputs(plot2_brush = list(xmin = 10, xmax = 2333, ymin = 40, ymax = 140),allowInputNoBinding_ = TRUE)
 app$setInputs(button_run_auto = "click")
 app$snapshot()
