+++ inst/mtr_app/tests/generate_series.R
@@ -1,5 +1,5 @@
 app <- ShinyDriver$new("../")
-app$snapshotInit("run_detection")
+app$snapshotInit("generate_series")
 
 app$setInputs(sidebarCollapsed = FALSE)
 # Input 'plot2_brush' was set, but doesn't have an input binding.
@@ -15,4 +15,6 @@ app$setInputs(buttoncreatpath2 = "click")
 # Input 'plot2_brush' was set, but doesn't have an input binding.
 app$setInputs(plot2_brush = list(xmin = 10, xmax = 2333, ymin = 60, ymax = 120),allowInputNoBinding_ = TRUE)
 app$setInputs(button_run_auto = "click")
+app$snapshot()
+app$setInputs(button_results = "click")
 app$snapshot()
