### 2019-03-27 9977cc5
rewrite some functions
