+++ R/Functions for plotting.R
@@ -1,0 +1,210 @@
+# Functions for plotting detected ring borders and labels
+
+create.label <- function(bor.col, sample.yr, dp)
+{
+  year.number <- sample.yr:(sample.yr - length(bor.col) + 1)
+  border.number <- length(bor.col) %>% seq_len
+  rw <- diff(bor.col) / dp
+  year.rw <- year.number[-1]
+  dfrw <- data.frame(col.num = bor.col[-1], rw = rw, year = year.rw)
+  label.list <- list(year = year.number, sn = border.number, dfrw = dfrw)
+  return(label.list)
+}
+
+split.label <- function(dfrw, x.left, x.right, seg)
+{
+  bor.list<-list()
+  rw.list<-list()
+  yr.list<-list()
+  for (i in 1:seg) {
+    bor.range <- which(dfrw[, 1] >= x.left[i] & dfrw[, 1] <= x.right[i])
+    df.bor <- dfrw[, 1][bor.range]
+    df.rw <- dfrw[, 2][bor.range]
+    df.yr <- dfrw[, 3][bor.range]
+    bor.list <- c(bor.list, list(df.bor))
+    rw.list <- c(rw.list, list(df.rw))
+    yr.list <- c(yr.list, list(df.yr))
+  }
+  split.list <- list(bor.list = bor.list, 
+    rw.list = rw.list, 
+    yr.list = yr.list)
+  return(split.list)
+}
+
+f.rw.plot <- function(bor.list, rw.list, yr.list, 
+  x.left, x.right, seg.name, location)
+{
+  plot(bor.list, rw.list, xlim = c(x.left, x.right), 
+    ylim = c(round(min(rw.list), 1) - 0.3, 
+      round(max(rw.list), 1) + 0.3), 
+    main = "", xlab = "", ylab = "", type = "o", 
+    cex = 1, axes = F, lty = 3, pch = 16)
+  if (location == 'upper') {
+    axis(1, at = bor.list, labels = yr.list, 
+      las = 2, font = 3, cex.axis = 1.2)
+    title(main = seg.name, cex.main = 1.5, line = -1)
+  } else {
+    axis(3, at = bor.list, labels = yr.list, 
+      las = 2, font = 3, cex.axis = 1.2)
+  }
+  axis(2, at = seq(round(min(rw.list), 1) - 0.3, 
+    round(max(rw.list), 1) + 0.3, by = 0.4), 
+    cex.axis = 1, las = 2, line = -1)
+  title(ylab = "Ring Widths (mm)", cex.lab = 1.2, line = 2.8)
+}
+
+f.img <- function(ring.data, x.left, x.right, ybottom, ytop, x.left1, incline)
+{
+  if (incline) {
+    plot(0, 0, xlab = "", ylab = "",
+      xlim = c(x.left, x.right), 
+      ylim = c(ybottom - 0.25 * ytop, ybottom + 1.25 * ytop), 
+      axes = F, main = "", type = "n")
+  } else {
+    plot(0, 0, xlab = "", ylab = "",
+      xlim = c(x.left, x.right), 
+      ylim = c(ybottom, ybottom + 1.25 * ytop), 
+      axes = F, main = "",  type = "n")
+    axis(1, cex.axis = 1)
+  }
+  axis(2, cex.axis = 1, line = -1, at = c(ybottom, ybottom + ytop))
+  x.range <- (x.left-x.left1+1):(x.right-x.left1+1)
+  x1 <- ifelse(is.matrix(ring.data), 
+    assign('plot.data',ring.data[, x.range]), 
+    assign('plot.data',ring.data[, x.range, ]))
+  rasterImage(plot.data, x.left, ybottom, x.right,
+    ybottom + ytop, interpolate = TRUE)
+}
+
[…]
