### 2019-03-27 acfc7e0
Update test-imgInput.R
