### 2019-04-28 946d8ae
Fix typos and add spell check.
