+++ inst/mtr_app/app.R
@@ -22,15 +22,15 @@ createUI <- function()
   shiny.title <- dashboardHeader(title = 'MtreeRing')
   shiny.sider <- dashboardSidebar(
     sidebarMenu(
-      menuItem('Image Loading',tabName = 'input_pre', 
+      menuItem('Image Loading',tabName = 'input_pre',
         icon = icon('folder-open', lib = 'font-awesome')),
-      menuItem('Measurement',tabName = 'mea_arg', 
+      menuItem('Measurement',tabName = 'mea_arg',
         icon = icon('gear', lib = 'font-awesome'), selected = TRUE)
     )
   )
   page1 <- fluidRow(
     box(
-      title = div(style = 'color:#FFFFFF;font-size:80%; 
+      title = div(style = 'color:#FFFFFF;font-size:80%;
         font-weight: bolder', 'Image Preview'),
       width = 12, status = 'primary', solidHeader = T, collapsible = T,
       plotOutput('pre.img',
@@ -39,7 +39,7 @@ createUI <- function()
           opacity = 0.25,
           resetOnNew = TRUE)
       )
-      ), 
+      ),
     box(
       title = div(style = 'color:#FFFFFF;font-size:80%;
         font-weight: bolder', 'Image Upload'),
@@ -55,8 +55,8 @@ createUI <- function()
         ' formats include png, jpg, tif and bmp.',
         style = 'color:#000000;font-size:90%'),
       prettyCheckbox(
-        inputId = "inmethod", 
-        label = div(style = 'color:#000000;font-weight: bolder;','Image Path'), 
+        inputId = "inmethod",
+        label = div(style = 'color:#000000;font-weight: bolder;','Image Path'),
         shape = "curve", value = F, status = "success"),
       conditionalPanel(
         condition = 'input.inmethod',
@@ -108,20 +108,20 @@ createUI <- function()
         status = "danger", shape = "square",
         fill = FALSE, inline = FALSE),
       prettyCheckbox(
-        inputId = "showcropp", 
+        inputId = "showcropp",
         label = div(style = 'color:#000000;font-weight: bolder;', 'Show Help'),
         shape = "curve", value = F, status = "success"
       ),
       conditionalPanel(
         condition = 'input.showcropp',
         helpText(
-          "The operation \"brush\" allows users to create a transparent ", 
-          "rectangle on the image and drag it around. For cores scanned ", 
-          "side by side, the user can choose a core of interest by brushing.", 
+          "The operation \"brush\" allows users to create a transparent ",
+          "rectangle on the image and drag it around. For cores scanned ",
+          "side by side, the user can choose a core of interest by brushing.",
           style = 'color:#000000;text-align:justify;'),
         helpText(
           "After brushing, click on the button \"Crop\" to create a",
-          " cropped area. The measurement will be performed within", 
+          " cropped area. The measurement will be performed within",
           " this area, rather than the whole (uncropped) image.",
           style = 'color:#000000;text-align:justify;'),
         helpText(
@@ -130,7 +130,7 @@ createUI <- function()
           "click on the outer region of the rectangle (this will make the",
           " rectangle disappear) and then click on the button \"Cancel\".",
           style = 'color:#FF0000;text-align:justify;')
-      ),  
+      ),
       hr(),
       actionButton(
         'buttoncrop', 'Crop',
@@ -150,70 +150,70 @@ createUI <- function()
       textInput('sample_yr', 'Sampling year', '', '75%'),
[…]
