+++ R/autoDetect.R
@@ -2,144 +2,144 @@
 #' @importFrom dplyr filter group_by select summarize
 #' @importFrom measuRing graySmoothed linearDetect
 #' @importFrom spatstat connected im
-#' @importFrom imager as.cimg cannyEdges dilate_rect erode_rect imgradient 
+#' @importFrom imager as.cimg cannyEdges dilate_rect erode_rect imgradient
 #' mclosing_square threshold watershed
 #' @title Automatic detection of tree-ring boundaries
-#' @description This function is used to automatically detect tree-ring 
+#' @description This function is used to automatically detect tree-ring
 #' boundaries along the user-defined path.
 #' @author Jingning Shi
 #' @param ring.data A magick image object produced by \code{imgInput}.
 #' @param seg An integer specifying the number of image segments.
-#' @param auto.path A logical value. If \code{TRUE}, a path is automatically 
-#' created at the center of the image. If \code{FALSE}, the function allows 
-#' the user to create a sub-image and a path by interactive clickings. 
+#' @param auto.path A logical value. If \code{TRUE}, a path is automatically
+#' created at the center of the image. If \code{FALSE}, the function allows
+#' the user to create a sub-image and a path by interactive clickings.
 #' See details below.
-#' @param manual A logical value indicating whether to skip the automatic 
-#' detection. If \code{TRUE}, ring boundaries are visually identified using 
+#' @param manual A logical value indicating whether to skip the automatic
+#' detection. If \code{TRUE}, ring boundaries are visually identified using
 #' the function \code{visualSelect}.
-#' @param method A character string specifying how ring borders are detected. 
-#' It requires one of the following characters: \code{"watershed"}, 
+#' @param method A character string specifying how ring borders are detected.
+#' It requires one of the following characters: \code{"watershed"},
 #' \code{"canny"}, or \code{"lineardetect"}. See details below.
-#' @param incline A logical value indicating whether to correct ring widths. 
+#' @param incline A logical value indicating whether to correct ring widths.
 #' If \code{TRUE}, two horizontal paths are added to the image.
-#' @param sample.yr \code{NULL} or an integer giving the year of formation 
+#' @param sample.yr \code{NULL} or an integer giving the year of formation
 #' of the left-most ring. If \code{NULL}, use the current year.
-#' @param watershed.threshold The threshold used for producing the marker 
-#' image, either a numeric from 0 to 1, or the character "auto" (using the 
+#' @param watershed.threshold The threshold used for producing the marker
+#' image, either a numeric from 0 to 1, or the character "auto" (using the
 #' Otsu algorithm), or a character of the form "XX\%" (e.g., "58\%").
-#' @param watershed.adjust A numeric used to adjust the Otsu threshold. 
-#' The default is 1 which means that the threshold will not be adjusted. 
-#' The sizes of early-wood regions in the marker image will reduce along 
+#' @param watershed.adjust A numeric used to adjust the Otsu threshold.
+#' The default is 1 which means that the threshold will not be adjusted.
+#' The sizes of early-wood regions in the marker image will reduce along
 #' with the decrease of \code{watershed.adjust}.
-#' @param struc.ele1 \code{NULL} or a vector of length two specifying the 
-#' width and height of the first structuring element. If \code{NULL}, the 
+#' @param struc.ele1 \code{NULL} or a vector of length two specifying the
+#' width and height of the first structuring element. If \code{NULL}, the
 #' size of the structuring element is determined by the argument \code{dpi}.
[…]

[…]
