### 2019-04-28 #287 @rorynolan
Hello @JingningShi, @annakrystalli and @bhive01
Here is my initial review.

@JingningShi, great job on a really nice package. I have several improvements that I'd like you to implement but I like the package a lot, it was nice to try it out.

---

* Take installation instructions out of vignette, they belong in the README only.
* Vignettes should be at least as extensive as the README. It's OK to duplicate examples from the README, most people will read one or the other.
* You need 2 vignettes: one for how to use the package with shiny and one for how to use from R console without shiny. Apart from `imgInput()`, there are only 4 exported functions: `autoDetect()`, `calcRingWidth()`, `nearPith()` and `visualSelect()`. All of these should be demonstrated with a video or gif in that vignette (you could demonstrate more than one function per video/gif e.g. `autoDetect()` and `calcRingWidth()` could be demonstrated in the same video/gif if you like). I know it seems like a lot of effort to demonstrate both the shiny and R console use because they're similar but most people will use one or the other and will appreciate a video/gif showing how to do it _exactly_ the way they're trying to do it.

* Why allow `magick = FALSE`, why not just insist on the use of `magick`?
* I think you should include an option not to plot the image when reading it. Some people would prefer the image to display for the first time when another function e.g. `nearPith()` is called.

* Your tests produce a lot of plots in `tests/testthat`. Your tests should either not produce these or remove them; they shouldn't still be there when the tests have finished (for some reason, when I run `devtools::test()` the plots are produced but not when I run `devtools::check()`).

#### Continuous Integration

* Can you add AppVeyor CI? If not, why not? It has helped me catch bugs in a few of my packages.

* Why is `nearPith()` named like this? Is it because inner arcs might be used?
* When you mention one function in the documentation of another, there should be a documentation link. For example, the mention of `visualSelect()` in `nearPith()` should be a clickable link.

#### Typos

* I've submitted a PR (https://github.com/JingningShi/MtreeRing/pull/2) to fix a few typos in messages displayed to users and to include a spell check when testing.

#### rOpenSci packaging guidelines

See https://ropensci.github.io/dev_guide/.

* You need to run `codemetar::write_codemeta()`.
* Consider `snake_case` style and an object_verb() naming scheme for functions in your package. For example you could rename `imgInput()` to `img_input()`. This is a matter of personal taste though, I don't insist that you do this.
* Your repo should include a `README.Rmd` file which is used to build `README.md`. You should not edit `README.md` directly.
* Create a website for your package with the `pkgdown` package and have it deploy using travis with every push to github (again, see https://ropensci.github.io/dev_guide/).
