### 2019-04-28 #287 @bhive01
---

### Function Naming

You have quite the mix of function naming. Many functions use camelCase, a few use _ and a lot of them use '.'. https://ropensci.github.io/dev_guide/building.html#function-and-argument-naming suggests using snake_case which is used in a few instances.

> ls(getNamespace("MtreeRing"), all.names = TRUE)
 [1] ".__DEVTOOLS__"        ".__NAMESPACE__."      ".__S3MethodsTable__." ".packageName"         "add_path"
 [6] "autoDetect"           "bor.distance"         "calcRingWidth"        "columnIndices"        "convert2gray"
[11] "correct.color"        "create_path"          "create.label"         "draw_border1"         "draw_border2"
[16] "f.bor.plot"           "f.border"             "f.img"                "f.marker"             "f.morphological"
[21] "f.plot.double"        "f.plot.single"        "f.rw.plot"            "f.sort"               "hat"
[26] "imgInput"             "inclined_path"        "launchMtRApp"         "magick2array"         "mark_arc"
[31] "nearPith"             "r.det"                "split.label"          "visualSelect"         "water.im"
[36] "watershed.im"

## autoDetect.R
method = "lineardetect" fails with example and there is no indication why.

`autoDetect(ring.data = t1, seg = 1, method = 'george')`, did also not throw an error. Some checking of the inputs here would go a long way for usability.

`auto.path = FALSE` fails
```
> autoDetect(ring.data = t1, seg = 1, auto.path = FALSE)
Error in mtext(text.s1, side = 1, line = text.line, adj = 0) :
  plot.new has not been called yet
```

`autoDetect(ring.data = t1, seg = 1, manual = TRUE)` Does not load visual selection tool.
Perhaps I am misunderstanding things. If so, perhaps the documentation needs adjustment?

sample.yr defaults to the current year
`sample.yr <- Sys.Date() %>% as.character %>% strtrim(4) %>% as.numeric`
I think this is ok, but I think perhaps it should come with a warning when left as `NULL`.

## calcRingWidth.R
Two things that bugged me when I tried to use this function:

1. `seriesID` isn't a bad name really, but it took me a while to realize it's just the name of the new column. Perhaps a better name could be given to it?
2. You should specify that your return unit is millimeters. The only indication of that is `dp <- x.dpi / 25.4` with `rw <- round(diff.col.num / dp, 2)`. Perhaps this is a known entity in the tree world? (I work on annuals).

## f.img.R

For some reason when using autoDetect, I would get an empty quartz window every time I used it. I suspect the reason is in this file, but I can't see it obviously.
I like the file type handling code here. Simple and straightforward.

## launchMtRApp.R

This is good. Really good. Makes a really intuitive interface to the whole process. The only issue I had with this was when I wanted to stop running this app R would not respond. This was outside of Rstudio (**gasp**). In R-gui Mac, it just gave me the spinning color wheel of death until I force quit. My bad, but is this a known thing. I haven't had time to look.

When I did the auto border search I got the following message:
"27 boreders were detected"
Boreders should be borders.

The app works pretty well and is intuitive. How do you save the output?

## nearPith.R

The first time it loaded up and I didn't see the instructions to click. I finally realized this and clicked, but when I wanted to click on the arcs it failed. `Error in xy.coords(x, y) : 'x' and 'y' lengths differ`. I'm not sure what went wrong, but when I then tried to rerun the same code, it failed differently: `Error in mtext(paste0("Step ", step.number, ": Click the left ", "mouse button to add a horizontal path."),  :
  plot.new has not been called yet`
I reloaded R and reran it and I think I messed up. You click on the edges or start and stop points of the image first. Then you click all the borders of the rings. I stopped it and had to restart R again.
I reloaded again and redid it. Seemed to be ok. It marked the middle of the path, but then I wasn't sure what to do. How do I mark that as input to the function? Or store that path for further computation? :/

## visualSelect.R

The function seems to do well at removing existing borders, but when I say add = TRUE, it doesn't. Is this meant to only be run in Rstudio or something?

## watershed.im.R
Seems to be a repository for functions. Perhaps this should be tools.R or something similar?

## Overall

I think this is a really good fit for ropensci, but I think that perhaps it needs a bit more work to clarify what's going on with the interactive bits. I'd say it's close, but not quite there. The shiny app is awesome and one of the first examples I've come across that provide a tool to make the process of doing this sort of thing for those less familiar with the command line.
