+++ README.md
@@ -9,7 +9,8 @@ MtreeRing
 **License:** [GPL3](https://cran.r-project.org/web/licenses/GPL-3)
 
 <!--pkg badges-->
-[![Build Status](https://travis-ci.org/JingningShi/MtreeRing.svg?branch=master)](https://travis-ci.org/JingningShi/MtreeRing.svg?branch=master)
+[![TravisCI Build Status](https://travis-ci.org/JingningShi/MtreeRing.svg?branch=master)](https://travis-ci.org/JingningShi/MtreeRing.svg?branch=master)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/3iwxn4euvmw2fd6e/branch/master?svg=true)](https://ci.appveyor.com/project/JingningShi/mtreering/branch/master)
 [![codecov](https://codecov.io/gh/JingningShi/MtreeRing/branch/master/graphs/badge.svg)](https://codecov.io/gh/JingningShi/MtreeRing)
 [![](https://badges.ropensci.org/284_status.svg)](https://github.com/ropensci/software-review/issues/287)
 [![cran checks](https://cranchecks.info/badges/worst/MtreeRing)](https://cranchecks.info/pkgs/MtreeRing)
