+++ vignettes/MtreeRing_vignette.Rmd
@@ -102,7 +102,7 @@ If you are not familiar with R and its command line interface, the shiny-based a
 
 A workflow for the Shiny app can be found in the README file of this package. Most steps are demonstrated with a gif to make the workflow more understandable. Here is a link: https://github.com/JingningShi/MtreeRing.
 
-```{r eval=FALSE}
+```{r }
 launchMtRApp()
 ```
 
