+++ R/ring_read.R
@@ -1,0 +1,120 @@
+#' @title Read and plot a tree-ring image file
+#' @export
+#' @importFrom png readPNG
+#' @importFrom jpeg readJPEG
+#' @importFrom tiff readTIFF
+#' @importFrom bmp read.bmp
+#' @importFrom magrittr "%>%"
+#' @importFrom magick image_crop image_info image_read image_convert 
+#' image_rotate image_resize
+#' @importFrom grDevices as.raster dev.cur dev.new dev.set dev.off dev.list
+#' @importFrom graphics abline axis layout lines locator mtext par plot points 
+#' rasterImage segments text title
+#' @description This function can read an image file from the hard disk and 
+#' plot it in a newly-opened graphical device.
+#' @author Jingning Shi
+#' @param img A character string indicating the path of the image file. 
+#' Supported formats include png, tiff, jpg and bmp.
+#' @param dpi An integer specifying the dpi of the image file. A minimum of 
+#' 300 dpi is required when running automatic detection.
+#' @param RGB A numeric vector of length 3 giving the weight of RGB channels.
+#' @param rotate An integer specifying how many degrees to rotate (clockwise). 
+#' It requires one of the following values:
+#' \code{0}, \code{90}, \code{180} or \code{270}.
+#' @param magick A logical value. If \code{TRUE}, the package \code{magick} 
+#' is used to read the image file whose size is over 10MB.
+#' @return A magick image object containing the image data.
+#' @details 
+#' Proper image preparation has a great influence on the measurement of 
+#' ring widths. A tree-ring image should not contain irrelevant or redundant 
+#' features, such as wooden mounts where cores are glued. The larger the file 
+#' size of an image, the slower the image processing operation will be.
+#' 
+#' \bold{Pith side} of the sample should be placed on the \bold{right side} of 
+#' the graphical window. Use the argument \code{rotate} to change its position.
+#' 
+#' It is highly recommended to use the default value \code{magick = TRUE}, 
+#' because the package \code{magick} can significantly reduce the memory usage.
+#' @examples
+#' img.path <- system.file("001.png", package = "MtreeRing")
+#' 
+#' ## Read and plot the image:
+#' t1 <- ring_read(img = img.path, dpi = 1200)
+
+ring_read <- function(img, dpi = NULL, RGB = c(0.299, 0.587, 0.114), 
+                      rotate = 0, magick = TRUE)
+{
+  check.degree <- rotate %in% c(0, 90, 180, 270, 360) 
+  if (!check.degree)
+    stop(paste('The argument "rotate" should be one of',
+                 'the following integers: 0, 90, 180 or 270'))
+  if (is.null(dpi))
+    stop('Please provide the dpi value of the image')
+  options(warn = -1)
+  img.size <- file.size(img) / 1024^2
+  if (img.size <=10 | !magick) {
+    last.path <- basename(img)
+    if (grepl("\\.tif", last.path))
+      tree.data <- readTIFF(img, native = FALSE, info = FALSE)
+    if (grepl("\\.png", last.path))
+      tree.data <- readPNG(img, native = FALSE, info = FALSE)
+    if (grepl("\\.jpg", last.path) | grepl("\\.jpeg", last.path))
+      tree.data <- readJPEG(img, native = FALSE)
+    if (grepl("\\.bmp", last.path))
+      tree.data <- read.bmp(img) / 255
+    td.dim <- dim(tree.data)
[…]

[…]
