+++ man/ring_app_launch.Rd
@@ -15,11 +15,13 @@ Run a Shiny-based application within the system's default
 web browser.
 }
 \details{
-A workflow for the Shiny app can be found in the README file of this 
-package. Here is a link \url{https://github.com/JingningShi/MtreeRing}
+A workflow for the Shiny app can be found in the package vignette. 
+Type \code{vignette('app-MtreeRing')} to learn how to use the app. Most
+steps are demonstrated with a gif to make the workflow more understandable.
 
-To stop the app, press the Escape key, or click the stop sign icon 
-in the upper right corner of the RStudio console.
+To stop the app, go to the R console and press the Escape key. 
+You can also click the stop sign icon in the
+upper right corner of the RStudio console.
 }
 \author{
 Jingning Shi, Wei Xiang
