+++ _pkgdown.yml
@@ -1,3 +1,6 @@
+title: MtreeRing
+url: https://github.com/JingningShi/MtreeRing
+
 navbar:
   structure:
     left:
