+++ README.md
@@ -64,7 +64,7 @@ t2 <- ring_detect(ring.data = t1, seg = 2, method = 'watershed')
 <center><img src="vignettes/img/README-img001.png" width = 65% height = 65% /></center>
 <center>Figure 1. The automatic detection of ring borders</center>
 
-### 4. Calculate ring-width series 
+### 3. Calculate ring-width series 
 
 If all ring borders are correctly identified, you can generate a ring-width series in data frame format. Use `write.rwl` to export the ring-width series to an rwl file.
 
