### 2019-05-08 dc9b7ad
Update test-pith_measure.R
