### 2019-05-08 3446b1a
Update test-pith_measure.R
