### 2019-05-09 775059b
Update ring_detect.R
