### 2019-05-09 48ef42e
Update test-ring_detect.R
