### 2019-05-09 bde9c78
Update ring_detect.R
