+++ tests/testthat/test-ring_detect.R
@@ -5,54 +5,64 @@ test_that("ring_detect returns an array", {
   
   library(mockery)
   
+  path0 <- system.file("001.png", package = "MtreeRing")
+  img0 <- ring_read(img = path0, dpi = 1200, plot = T)
+  t0 <- ring_detect(ring.data = img0, sample.yr = 2015, method = 'watershed')
+  dev.off(as.numeric(attributes(img0)$dn))
+  dn <- attributes(t0)$seg.dn
+  apply(matrix(dn, nrow = 1), 2, dev.off)
+  
   path1 <- system.file("001.png", package = "MtreeRing")
-  img1 <- ring_read(img = path1, dpi = 1200, plot = FALSE)
-  t1 <- ring_detect(ring.data = img1, auto.path = T, method = 'watershed')
+  img1 <- ring_read(img = path1, dpi = 1200)
+  t1 <- ring_detect(ring.data = img1, sample.yr = 2015, method = 'watershed')
   dn <- attributes(t1)$seg.dn
   apply(matrix(dn, nrow = 1), 2, dev.off)
   
   img1 <- ring_read(img = path1, dpi = 1200)
-  dev.off(attributes(img1)$dn)
-  t1 <- ring_detect(ring.data = img1, auto.path = T, method = 'watershed')
+  t1 <- ring_detect(ring.data = img1, sample.yr = 2015, method = 'watershed')
   dn <- attributes(t1)$seg.dn
   apply(matrix(dn, nrow = 1), 2, dev.off)  
-  t2 <- ring_detect(ring.data = img1, seg = 3, auto.path = T, method = 'canny')
+  
+  t2 <- ring_detect(img1, seg = 3, sample.yr = 2015, method = 'canny')
   dn <- attributes(t2)$seg.dn
-  apply(matrix(dn, nrow = 1), 2, dev.off)  
-  t21 <- ring_detect(ring.data = img1, seg = 3, auto.path = T, 
+  apply(matrix(dn, nrow = 1), 2, dev.off) 
+  
+  t21 <- ring_detect(ring.data = img1, seg = 3, sample.yr = 2015, 
     method = 'canny', canny.smoothing = 1, 
     canny.t1 = 0.3, canny.t2 = 0.2)
   dn <- attributes(t21)$seg.dn
-  apply(matrix(dn, nrow = 1), 2, dev.off)  
-  t3 <- ring_detect(ring.data = img1, auto.path = T, method = 'lineardetect')
+  apply(matrix(dn, nrow = 1), 2, dev.off)
+  
+  t3 <- ring_detect(img1, sample.yr = 2015, method = 'lineardetect')
   dn <- attributes(t3)$seg.dn
   apply(matrix(dn, nrow = 1), 2, dev.off)  
-  t4 <- ring_detect(ring.data = img1, auto.path = T, manual = T)
-  dn <- attributes(t4)$seg.dn
-  apply(matrix(dn, nrow = 1), 2, dev.off)  
-  t5 <- ring_detect(ring.data = img1, auto.path = T, manual = T, incline = T)
-  dn <- attributes(t5)$seg.dn
-  apply(matrix(dn, nrow = 1), 2, dev.off)
+  
+  # t4 <- ring_detect(ring.data = img1, sample.yr = 2015, manual = F)
+  # dn <- attributes(t4)$seg.dn
+  # apply(matrix(dn, nrow = 1), 2, dev.off)  
+  
+  # t5 <- ring_detect( img1, sample.yr = 2015, manual = T, incline = T)
+  # dn <- attributes(t5)$seg.dn
+  # apply(matrix(dn, nrow = 1), 2, dev.off)
   
   
[…]
