+++ tests/testthat/test-create_path.R
@@ -5,7 +5,7 @@ test_that("mock test", {
   library(mockery)
   
   path1 <- system.file("001.png", package = "MtreeRing")
-  img1 <- ring_read(img = path1, dpi = 1200)
+  img1 <- ring_read(img = path1, dpi = 1200, plot = T)
   mock1 <- mock(c(50, 1800), cycle = T)
   mock2 <- mock(c(20, 160), cycle = T)
   mock3 <- mock(list(y = 111), cycle = T)
@@ -21,8 +21,8 @@ test_that("mock test", {
   
   p3 <- create_path(F, 'watershed', F, 1, 40, 2300, 170, 'black') 
   expect_is(p2, 'numeric')
-  
   dev.off(attributes(img1)$dn)
   
+
 })
 

[…]
