+++ vignettes/detection-MtreeRing.Rmd
@@ -35,7 +35,7 @@ t1 <- ring_read(img = img.name, dpi = 1200, plot = TRUE)
 t2 <- ring_detect(ring.data = t1, auto.path = TRUE, seg = 2, method = 'watershed')
 ```
 
-If `auto.path = TRUE`, the function will create a path at the center of the image. In the current version, the path is a horizontal dashed line. Detected ring borders are placed along the path, and are tagged with years and border numbers.
+If `auto.path = TRUE`, `ring_detect` will create a path at the center of the image. In the current version, the path is a horizontal dashed line. Detected ring borders are placed along the path, and are tagged with years and border numbers.
 
 <img src="img/single_path.png" width = 85% height = 85% />  
 
@@ -66,11 +66,23 @@ t2 <- ring_modify(ring.data = t2, del = c(1:4))
 
 ### 3.2 Add tree rings
 
-We have just removed 4 ring borders, now re-add these borders by left-clicking on the image. See function's help for more details about mouse clicks.
+We have just removed 4 ring borders, now re-add these borders by left-clicking on the image.
 
 ```{r eval=FALSE}
 t2 <- ring_modify(ring.data = t2, add = TRUE)
 ```
+
+If `add = TRUE`, graphics windows opened by `ring_detect` will be activated sequentially. When a graphics window is activated, you can add new borders by left-clicking the mouse along the path. Every click draws a point representing the ring border. The identification process does not automatically stop by itself.
+
+  1. On the Windows system, the identification process can be terminated by pressing the right mouse button and selecting **Stop** from the menu.
+  
+  2. On the MacOS system, for a X11 device the identification process is terminated by pressing any mouse button other than the first, and for a quartz device this process is terminated by pressing the **ESC** key.
+  
+Once you terminate an identification process, the current graphics window will be closed automatically, and the graphics window of the following segment is activated.
+
+Here is an example of adding ring borders on Windows system:
+
+<img src="img/Ring_modify_add.gif" width = 95% height = 95% />
 
 ## 4. Calculate ring-width series 
 
