### 2019-05-11 #287 @JingningShi
Sorry for the delay in the response. Thank you again for taking the time to thoroughly review the package @rorynolan @bhive01 . English isn't my first language, so please excuse any mistakes. I will explain to the best of my ability.

I'll first respond to @rorynolan  comments:

## Vignettes

>  1. Take installation instructions out of vignette, they belong in the README only.

Installation instructions have been removed.

>  2. Vignettes should be at least as extensive as the README. It's OK to duplicate examples from the README, most people will read one or the other. You need 2 vignettes....

Thanks for the excellent suggestion.  I deleted the old vignette and wrote two new vignettes for how to use these functions from the R console. Each vignette is demonstrated with pictures.

I also wrote the third vignette for how to use the shiny app. The third vignette is demonstrated with gifs, making the workflow of shiny app more understandable.

```{r eval=FALSE}
vignette('detection-MtreeRing')
vignette('pith-MtreeRing')
vignette('app-MtreeRing')
```

## Code

> 1.  Why allow magick = FALSE, why not just insist on the use of magick?

The major goal of `magick` is to provide basic image transformations, such as image cropping and rotation, rather than reading an image.

In some cases, the image data is stored in a non-standard format. When reading such a  file, we found that packages `png`, `jpg` and `tiff` exhibited better compatibility with non-standard file formats than `magick`. The former just throws up a warning which doesn't affect the image data, but `magick` returns an error and stops the execution of image reading. However, the storage mode of the image data obtained by `png`, `jpg` and `tiff` is `double`, which occupies substantial amounts of memory (e.g., an image file of twelve cores at a resolution of 2540 dpi in the png format occupies 72.8 MB of disk space, whereas the corresponding raster array created by function `readPNG` occupies approximately 3.8 GB of memory space). For this reason, we use `png`, `jpg` and `tiff` to read small-size image files.

To consume less memory, we use `magick` to read images whose file size is over 10MB. For large files stored in a non-standard format, we can set `magick = FALSE` to avoid throwing an error.

> 2. I think you should include an option not to plot the image when reading it. Some people would prefer the image to display for the first time when another function e.g. nearPith() is called.

This is a good suggestion. I am not thoughtful enough.

I have included an argument `plot` in `ring_read`. Now the user can choose not to display a tree-ring image until another function, such as `ring_detect`, is called.

## Tests

> Your tests produce a lot of plots in tests/testthat. Your tests should either not produce these or remove them; they shouldn't still be there when the tests have finished (for some reason, when I run devtools::test() the plots are produced but not when I run devtools::check()).

I'm not sure what went wrong. When I run `devtools::test`, I don't get those plots on my Windows or macOS computer.

## Continuous Integration

> Can you add AppVeyor CI? If not, why not? It has helped me catch bugs in a few of my packages.

This is a great suggestion. I add [a new CI test](https://ci.appveyor.com/project/JingningShi/mtreering/branch/master) using AppVeyor and this CI test passes.

## Documentation

> 1. Why is nearPith() named like this? Is it because inner arcs might be used?

Thanks for this very nice suggestion. I rename `nearPith` to `pith_measure`. Now, all functions in `MtreeRing` use an `object_verb()` naming scheme.

> 2. When you mention one function in the documentation of another, there should be a documentation link. For example, the mention of visualSelect() in nearPith() should be a clickable link.

All functions in the documentation are now clickable.

## rOpenSci packaging guidelines

> 1. You need to run codemetar::write_codemeta().

The codemeta file has been added.

> 2. Consider snake_case style and an object_verb() naming scheme for functions in your package. For example you could rename imgInput() to img_input(). This is a matter of personal taste though, I don't insist that you do this.

Thanks for this helpful suggestion. I rename all functions (including both internal functions and exported functions) using a snake_case style. Now, most exported functions start with  `ring_` and end with a `verb` which indicates the primary action of a function, such as `ring_read` and `ring_detect`.

| original function names |  new function names |
| ------ | ------ |
| imgInput | ring_read |
| autoDetect | ring_detect |
| visualSelect | ring_modify |
| calcRingWidth | ring_calculate |
| launchMtRApp | ring_app_launch |
| nearPith | pith_measure |

> 3. Your repo should include a README.Rmd file which is used to build README.md. You should not edit README.md directly.

The README.Rmd file has been added and listsed in .Rbuildignore.

> 4. Create a website for your package with the pkgdown package and have it deploy using travis with every push to github (again, see https://ropensci.github.io/dev_guide/).

Both HTML documentation and the _pkgdown.yml file have been added.

> 5. Run usethis::use_code_of_conduct() to add an appropriate code of conduct for potential contributors.

The CODE_OF_CONDUCT.md file has been added, and the README.md includes a link to the CODE_OF_CONDUCT.md file.

----

Response to @bhive01 comments

## Function Naming

> You have quite the mix of function naming. Many functions use camelCase, a few use _ and a lot of them use '.'. https://ropensci.github.io/dev_guide/building.html#function-and-argument-naming suggests using snake_case which is used in a few instances.

Thanks for this helpful suggestion. I rename these functions (including both internal functions and exported functions) using a snake_case style. Now, most exported functions start with  `ring_` and end with a `verb` which indicates the primary action of a function, such as `ring_read` and `ring_detect`.

| original function names |  new function names |
| ------ | ------ |
| imgInput | ring_read |
| autoDetect | ring_detect |
| visualSelect | ring_modify |
| calcRingWidth | ring_calculate |
| launchMtRApp | ring_app_launch |
| nearPith | pith_measure |

> ls(getNamespace("MtreeRing"), all.names = TRUE)
 [1] ".__NAMESPACE__."      ".__S3MethodsTable__." ".packageName"
 [4] "add_path"             "bor_distance"         "border_det"
 [7] "border_plot"          "color_correct"        "column_indices"
[10] "conn_calc"            "convert2gray"         "create_path"
[13] "draw_border1"         "draw_border2"         "hat"
[16] "img_plot"             "inclined_path"        "label_create"
[19] "label_split"          "magick2array"         "mark_arc"
[22] "marker_plot"          "morphology_process"   "pith_measure"
[25] "ring_app_launch"      "ring_calculate"       "ring_det"
[28] "ring_detect"          "ring_modify"          "ring_read"
[31] "ring_sort"            "rw_plot"              "single_path_plot"
[34] "two_paths_plot"       "watershed_process"

## autoDetect.R

> 1. method = "lineardetect" fails with example and there is no indication why.

Sorry about that. This bug has been fixed. The following calls can now properly detect ring borders.
```
img.path <- system.file("001.png", package = "MtreeRing")
t1 <- ring_read(img = img.path, dpi = 1200, plot = F)
t2 <- ring_detect(t1, seg = 1, method = 'lineardetect', border.color = 'green')
```
> 2. autoDetect(ring.data = t1, seg = 1, method = 'george'), did also not throw an error. Some checking of the inputs here would go a long way for usability.

Now `ring_detect` will check important inputs before detecting ring borders.

> 3. auto.path = FALSE fails

When `auto.path = FALSE`,  a user needs to create a path by clicking on the tree ring image produced by `ring_read`. If you close this graphics window before running `ring_detect`, you will get this error.

To avoid that, I have included an argument `plot` in `ring_read`. If `plot = FALSE` (the default), `ring_read` will not plot the tree-ring image until another function, such as `ring_detect`, is called.

The path creation demonstrated with pictures can be found in this vignette
```
vignette('detection-MtreeRing')
```

> 4. autoDetect(ring.data = t1, seg = 1, manual = TRUE) Does not load visual selection tool.
Perhaps I am misunderstanding things. If so, perhaps the documentation needs adjustment?

Thank you for this thoughtful comment, it greatly improved the user-friendliness of tree ring measurements. In a modified version, if `manual = TRUE`, `ring_detect` will load `ring_modify` after creating the path and the user could mark ring borders manually.

> 5. sample.yr defaults to the current year
sample.yr <- Sys.Date() %>% as.character %>% strtrim(4) %>% as.numeric
I think this is ok, but I think perhaps it should come with a warning when left as NULL.

If  `sample.yr = NULL`,  `ring_read` will now generate a warning message like this
```
 Warning message:
The sampling year is set to the current year
```

## calcRingWidth.R

> 1. seriesID isn't a bad name really, but it took me a while to realize it's just the name of the new column. Perhaps a better name could be given to it?

`series ID` is already used in a dendrochronological package `dplR`. It is defined as the column names of tree ring datasets. To ensure a consistent style across packages, I name this argument like this.

> 2. You should specify that your return unit is millimeters. The only indication of that is dp <- x.dpi / 25.4 with rw <- round(diff.col.num / dp, 2). Perhaps this is a known entity in the tree world? (I work on annuals).

Thanks for this useful suggestion! The unit of ring widths is now described in the function documentation.

## f.img.R

> 1. For some reason when using autoDetect, I would get an empty quartz window every time I used it. I suspect the reason is in this file, but I can't see it obviously.

I couldn't reproduce that error on my machine. When I call `ring_detect` in both RStudio and R GUI on a MacOS X computer, I do not get an empty quartz window.  Could you please provide more information?

## launchMtRApp.R

> 1. This is good. Really good. Makes a really intuitive interface to the whole process. The only issue I had with this was when I wanted to stop running this app R would not respond. This was outside of Rstudio (gasp). In R-gui Mac, it just gave me the spinning color wheel of death until I force quit. My bad, but is this a known thing. I haven't had time to look.

My mistake. The function documentation doesn't provide a clear and detailed description of how to stop this app.

To stop the app, you could go to the R console and press the Escape key. This works in both R GUI and RStudio. You could also click the stop sign icon in the upper right corner of the RStudio console.

This instruction has been added to the function documentation.

> 2. When I did the auto border search I got the following message:
"27 boreders were detected"
Boreders should be borders.

This spelling mistake has been corrected.

> 3. The app works pretty well and is intuitive. How do you save the output?

The output of tree ring measurements can be saved in either CSV files or RWL files (a file format used in tree ring analysis). I write a vignette for how to use the shiny app. Data saving can be found in the fifth part of this vignette.
```
vignette('app-MtreeRing')
```

## nearPith.R

> 1. The first time it loaded up and I didn't see the instructions to click. I finally realized this and clicked, but when I wanted to click on the arcs it failed. Error in xy.coords(x, y) : 'x' and 'y' lengths differ. I'm not sure what went wrong, but when I then tried to rerun the same code, it failed differently: Error in mtext(paste0("Step ", step.number, ": Click the left ", "mouse button to add a horizontal path."), : plot.new has not been called yet.
I reloaded R and reran it and I think I messed up. You click on the edges or start and stop points of the image first. Then you click all the borders of the rings. I stopped it and had to restart R again.
I reloaded again and redid it. Seemed to be ok. It marked the middle of the path, but then I wasn't sure what to do. How do I mark that as input to the function? Or store that path for further computation? :/

I reproduced these two errors on my computer and figured out why these errors happened. It‘s my fault. The original function documentation isn't very clear so I write a vignette for `pith_measure`. A detailed example is demonstrated with pictures in that vignette. I hope it is helpful.

```
vignette('pith-MtreeRing')
```

When you call `pith_measure`, it waits for the user to click on the tree ring image. `pith_measure` uses `locator` to record the position of the mouse cursor, and will call it many times. The `n` argument of `locator` controls the number of points to record in each call. During the arc selection, `locator` is used to record positions of arc endpoints.

```
arc.a <- locator(n = 1, type = "n")
points(arc.a$x, py, pch = border.type, col = color, cex = label.cex)
arc.b <- locator(n = 1, type = "n")
points(arc.b$x, py, pch = border.type, col = color, cex = label.cex)
```

If this process is terminated prematurely before clicking once on the arc (i.e., press the Escape key on MacOS computer, or click on ‘Stop’ button from the menu of the graphics device on Windows PC), `locator` will assign a `null` object to the variable `arc.a` or `arc.b`. In this case, a call to `points` results in the first error.

The second error occurs because the graphics device produced by `ring_read` is shut down and therefore `mtext` cannot write text onto the tree ring image. In a modified version, by default, `ring_read` will not open a graphics device until `pith_measure` is called by a user.

## visualSelect.R
> 1. The function seems to do well at removing existing borders, but when I say add = TRUE, it doesn't. Is this meant to only be run in Rstudio or something?

Similar to `pith_measure`, `ring_modify` uses `locator` to record the positions of points you added. `locator` is supported by both R GUI and RStudio.

When you call `ring_modify` with `add = TRUE`, graphics devices opened by `ring_detect` will be activated sequentially using `dev.set`. When a graphics device is activated (device name in the top right corner of the window is tagged with `ACTIVE`), it waits for the user to click on the image. After adding all borders in a device, you can terminate this process (e.g., press the Escape key on MacOS computer) and then add borders in another device.

 A complete procedure demonstrated with pictures can be found in the vignette
```
vignette('detection-MtreeRing')
```

## watershed.im.R
> 1. Seems to be a repository for functions. Perhaps this should be tools.R or something similar?

Thanks for pointing this out. I've now renamed `watershed.im` to `watershed_process` as suggested by [Packaging Guide](https://ropensci.github.io/dev_guide/building.html#function-and-argument-naming).
