### 2019-05-12 c6f42b9
update app documentation
