+++ man/ring_app_launch.Rd
@@ -7,14 +7,20 @@
 ring_app_launch(launch.browser = TRUE)
 }
 \arguments{
-\item{launch.browser}{A logical value. If \code{TRUE}, the system's default 
-web browser will be launched automatically after the app is started.}
+\item{launch.browser}{A logical value. 
+If \code{FALSE}, a built-in browser will be launched automatically 
+after the app is started. If \code{TRUE}, the system's default 
+web browser is used instead. This argument only works for RStudio.
+See details below.}
 }
 \description{
 Run a Shiny-based application within the system's default 
 web browser.
 }
 \details{
+\code{launch.browser = FALSE} is not recommended, as the file renaming
+does not work on the RStudio built-in browser when saving the data.
+
 A workflow for the Shiny app can be found in the package vignette. 
 Type \code{vignette('app-MtreeRing')} to learn how to use the app. Most
 steps are demonstrated with a gif to make the workflow more understandable.
