### 2019-05-13 #287 @rorynolan
Hi @JingningShi,
I'm very happy with your response to my comments. One last thing to me would be to include your above explanation of when one might not use `magick` in the documentation of `ring_read()`. I think it would be nice for users to have this clarified for them.
Other than that, I'm happy to sign off.
One last word of advice: be prepared for 1-2 years of basic maintenance of your package as users uncover bugs and make reasonable feature requests. This is quite a nice process really, not something to worry about.
