### 2019-05-14 34b9a31
update ring_read documentation
