### 2019-05-14 6cdb9a0
update ring_read documentation
