### 2019-05-14 #287 @JingningShi
Thanks @rorynolan , an explanation has been added to the R documentation.

> It is highly recommended to use the default value magick = TRUE, because the package magick can significantly reduce the memory usage. In a few cases where image data is stored in a non-standard format, magick may return an error when reading image files. In this case, you can set magick = FALSE to avoid the use of magick.

I will do my best to maintain and improve this package :smile:
