### 2019-05-14 7a9c78c
update R documentation
