+++ R/ring_read.R
@@ -24,8 +24,10 @@
 #' @param rotate An integer specifying how many degrees to rotate (clockwise). 
 #' It requires one of the following values:
 #' \code{0}, \code{90}, \code{180} or \code{270}.
-#' @param magick A logical value. If \code{TRUE}, the package \code{magick} 
-#' is used to read the image file whose size is over 10MB. See details below.
+#' @param magick A logical value. If \code{TRUE}, \code{magick} is used to
+#' read the image file whose size is over 10MB. If \code{FALSE},
+#' packages \code{png}, \code{jpg} and \code{tiff} are used instead.
+#' See details below.
 #' @return A magick image object containing the image data.
 #' @details 
 #' Proper image preparation has a great influence on the measurement of 
@@ -37,9 +39,11 @@
 #' of a graphics window. Use \code{rotate} to change its position.
 #' 
 #' It is highly recommended to use the default value \code{magick = TRUE}, 
-#' because the package \code{magick} can significantly reduce the memory usage.
+#' because the package \code{magick} can significantly reduce the memory usage
+#' when reading a large file.
 #' In a few cases where image data is stored in a non-standard format, 
-#' \code{magick} may return an error when reading image files. In this case,
+#' \code{magick} may return an error when reading files. 
+#' If \code{ring_read} fails reading a large file,
 #' you can set \code{magick = FALSE} to avoid the use of \code{magick}.
 #' 
 #' @examples
