+++ man/ring_read.Rd
@@ -24,8 +24,10 @@ function \code{\link{ring_detect}} or \code{\link{pith_measure}} is called.}
 It requires one of the following values:
 \code{0}, \code{90}, \code{180} or \code{270}.}
 
-\item{magick}{A logical value. If \code{TRUE}, the package \code{magick} 
-is used to read the image file whose size is over 10MB. See details below.}
+\item{magick}{A logical value. If \code{TRUE}, \code{magick} is used to
+read the image file whose size is over 10MB. If \code{FALSE},
+packages \code{png}, \code{jpg} and \code{tiff} are used instead.
+See details below.}
 }
 \value{
 A magick image object containing the image data.
@@ -44,9 +46,11 @@ size of an image, the slower the image processing operation will be.
 of a graphics window. Use \code{rotate} to change its position.
 
 It is highly recommended to use the default value \code{magick = TRUE}, 
-because the package \code{magick} can significantly reduce the memory usage.
+because the package \code{magick} can significantly reduce the memory usage
+when reading a large file.
 In a few cases where image data is stored in a non-standard format, 
-\code{magick} may return an error when reading image files. In this case,
+\code{magick} may return an error when reading files. 
+If \code{ring_read} fails reading a large file,
 you can set \code{magick = FALSE} to avoid the use of \code{magick}.
 }
 \examples{
