### 2019-05-14 #287 @bhive01
Going through this today.

👏🏻👏🏻👏🏻
Well done on the vignettes, examples, and function names. It's really good and I've run through them all without major issues. The gifs are super easy to follow and things are working better for me as a "naïve" user.

I'm happy to say that this package is a go for me now.
Really well done. Thanks for your efforts Jingning Shi.
