### 2019-05-19 #287 @annakrystalli
Actually, sorry @JingningShi, I just picked up that the package is missing a `NEWS.md` and that the package is still at version 1.2.

Tracking changes to the package is a requirement for an rOpenSci package. I also would expect that given the development and many improvements made to the package during review for the post-review package to have a higher version.  The `NEWS.md` is a great place to document all the changes and new features you have added.

Have a look at the chapter on  [releasing packages](https://ropensci.github.io/dev_guide/releasing.html#versioning) for more details and just reach out if you need any more help with this. Almost there!
