+++ README.Rmd
@@ -17,9 +17,9 @@ knitr::opts_chunk$set(
 **License:** [GPL3](https://cran.r-project.org/web/licenses/GPL-3)
 
 <!--pkg badges-->
-[![TravisCI Build Status](https://travis-ci.org/JingningShi/MtreeRing.svg?branch=master)](https://travis-ci.org/JingningShi/MtreeRing.svg?branch=master)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/3iwxn4euvmw2fd6e/branch/master?svg=true)](https://ci.appveyor.com/project/JingningShi/mtreering/branch/master)
-[![codecov](https://codecov.io/gh/JingningShi/MtreeRing/branch/master/graphs/badge.svg)](https://codecov.io/gh/JingningShi/MtreeRing)
+[![TravisCI Build Status](https://travis-ci.org/ropensci/MtreeRing.svg?branch=master)](https://travis-ci.org/ropensci/MtreeRing.svg?branch=master)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/64b6cgrdfjphm506/branch/master?svg=true)](https://ci.appveyor.com/project/ropensci/MtreeRing/branch/master)
+[![codecov](https://codecov.io/gh/ropensci/MtreeRing/branch/master/graphs/badge.svg)](https://codecov.io/gh/ropensci/MtreeRing)
 [![](https://badges.ropensci.org/287_status.svg)](https://github.com/ropensci/software-review/issues/287)
 [![cran checks](https://cranchecks.info/badges/worst/MtreeRing)](https://cranchecks.info/pkgs/MtreeRing)
 [![Downloads](https://cranlogs.r-pkg.org/badges/MtreeRing)](https://CRAN.R-project.org/package=MtreeRing)
@@ -114,4 +114,4 @@ The line segment connecting two dots on the same ring should match the tangent o
 
 Please note that the 'MtreeRing' project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
 
-
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
