+++ NEWS.md
@@ -1,0 +1,20 @@
+### MtreeRing 1.3 (2019-05-20)
+- Add a NEWS.md file to track changes
+- Add CODE_OF_CONDUCT.md
+- Add `plot` parameter to `ring_read`
+- Rename all exported functions using snake_case style
+- Update documentation for the Shiny app and mouse actions
+- Update README file
+- Fix a bug where `ring_detect` fails to detect ring borders
+- Remove intro vignette and add three new vignettes to demonstrate major functionality
+
+### MtreeRing 1.2 (2019-03-02)
+- Add appveyor-CI and travis-CI
+- Add README file
+- Add tests for all functions and Shiny app
+- Add intro vignette
+- Update documentation for `autoDetect`
+- Update DESCRIPTION URL and bug report
+
+### MtreeRing 1.1 (2018-10-16)
+- Initial CRAN release
