+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: MtreeRing
 Type: Package
 Title: A Shiny Application for Automatic Measurements of Tree-Ring Widths on Digital Images
-Version: 1.2
+Version: 1.3
 Authors@R: c(person("Jingning", "Shi", role = c("aut", "cre"), email = "snow940220@bjfu.edu.cn"), person("Wei", "Xiang", role = "aut"))
 Author: Jingning Shi [aut, cre], Wei Xiang [aut]
 Maintainer: Jingning Shi <snow940220@bjfu.edu.cn>
@@ -18,8 +18,8 @@ LazyData: TRUE
 NeedsCompilation: no
 Repository: CRAN
 Encoding: UTF-8
-URL: https://github.com/JingningShi/MtreeRing
-BugReports: https://github.com/JingningShi/MtreeRing/issues
+URL: https://github.com/ropensci/MtreeRing
+BugReports: https://github.com/ropensci/MtreeRing/issues
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
 Language: en-US
